==> code/R/IMCData_analysis/01_read_transform_data.Rmd <==
---
title: "Read in steinbock data"
output: html_document
date: '2022-08-12'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data {#read-data}

This section describes how to read in single-cell data and images into `R`
**after** image processing and segmentation (see Section \@ref(processing)).

We use the [imcRtools](https://github.com/BodenmillerGroup/imcRtools) package to
read in single-cell data extracted using the `steinbock` framework or the IMC
Segmentation Pipeline. Both image processing approaches also generate
multi-channel images and segmentation masks that can be read into `R` using the
[cytomapper](https://github.com/BodenmillerGroup/cytomapper) package.

```{r, message=FALSE}
library(imcRtools)
library(cytomapper)
library(BiocParallel)
```

## Read in single-cell information

For single-cell data analysis in `R` the
[SingleCellExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html)
[@Amezquita2019] data container is commonly used within the Bioconductor
framework. It allows standardized access to (i) expression data, (ii) cellular
metadata (e.g. cell type), (iii) feature metadata (e.g. marker name) and (iv)
experiment-wide metadata. For an in-depth introduction to the `SingleCellExperiment`
container, please refer to the [SingleCellExperiment class](https://bioconductor.org/books/3.14/OSCA.intro/the-singlecellexperiment-class.html).

The [SpatialExperiment](https://bioconductor.org/packages/release/bioc/html/SingleCellExperiment.html)
class [@Righelli2021] is an extension of the `SingleCellExperiment` class. It
was developed to store spatial data in addition to single-cell data and an
extended introduction is accessible
[here](https://bioconductor.org/packages/release/bioc/vignettes/SpatialExperiment/inst/doc/SpatialExperiment.html).

To read in single-cell data generated by the `steinbock` framework or the IMC
Segmentation Pipeline, the `imcRtools` package provides the `read_steinbock` and
`read_cpout` function, respectively. By default, the data is read into a
`SpatialExperiment` object; however, data can be read in as a
`SingleCellExperiment` object by setting `return_as = "sce"`. All functions
presented in this book are applicable to both data containers.

### steinbock generated data

The [steinbock](https://github.com/BodenmillerGroup/steinbock) framework was
used to process and segment IMC raw data available.

The `read_steinbock` function provided by `imcRtools` can now be used to read in
`steinbock` generated data. For more information, please refer to
`?read_steinbock`.

```{r read-steinbock}
path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20230112_DIMR_Ilastik/PT/20230308_sc_data/simple_segmentation"

#spe_bm <- read_steinbock(path=file.path(path, "BM/20230308_sc_data/simple_segmentation"), image_file=NULL, panel_file = file.path(path, 'BM/20230308_sc_data/simple_segmentation/NB_Panel.csv')) 
#spe_pt <- read_steinbock(path=file.path(path, "PT/20230308_sc_data/simple_segmentation"), image_file=NULL, panel_file = file.path(path, 'PT/20230308_sc_data/simple_segmentation/NB_Panel.csv')) 
#spe <- read_steinbock(path=path, image_file=NULL, panel_file = file.path(path, 'NB_Panel.csv')) 
#spe <- cbind(spe_pt, spe_bm)

spe <- read_steinbock(path=path, image_file=NULL, panel_file = file.path(path, 'NB_Panel.csv')) 
spe
```

By default, single-cell data is read in as `SpatialExperiment` object. 
The summarized pixel intensities per channel and cell (here mean intensity) are
stored in the `counts` slot. Columns represent cells and rows represent channels.

```{r counts}
rowData(spe)
counts(spe)[1:44,1:9]
```

Metadata associated to individual cells are stored in the `colData` slot. After
initial image processing, these metadata include the numeric identifier (`ObjectNumber`),
the area, and morphological features of each cell. In addition, `sample_id` stores
the image name from which each cell was extracted and the width and height of the
corresponding images are stored.

```{r colData}
head(colData(spe))
typeof(colData(spe))
class(spe)
class(colData(spe))
nrow(colData(spe))
colData(spe)[1:5, 2:ncol(colData(spe))]
```

The main difference between the `SpatialExperiment` and the
`SingleCellExperiment` data container in the current setting is the way spatial
locations of all cells are stored. For the `SingleCellExperiment` container, the
locations are stored in the `colData` slot while the `SpatialExperiment`
container stores them in the `spatialCoords` slot:

```{r spatialCoords}
head(spatialCoords(spe))
```

Finally, metadata regarding the channels are stored in the `rowData` slot. This
information is extracted from the `panel.csv` file. Channels are ordered by
isotope mass and therefore match the channel order of the multi-channel images
(see Section \@ref(read-images)).

```{r rowData}
head(rowData(spe))
rowData(spe)
```


## Single-cell processing {#cell-processing}

After reading in the single-cell data, few further processing steps need to be
taken.

**Add additional metadata**

We can set the `colnames` of the object to generate unique identifiers per cell:

```{r set-colnames}
colnames(spe) <- paste0(spe$sample_id, "_", spe$ObjectNumber)
```

It is also often the case that sample-specific metadata are available externally.
For the current data, we need to link metadata from an external excel file.
More information on how to extract matching patterns from a string can be found [here](https://stringr.tidyverse.org/reference/str_extract.html).

```{r add-indication}
library(stringr)

spe$sample <- gsub(spe$sample_id, pattern="([UM]).*", replacement="\\1")
spe$fm_id <- str_match(spe$sample, "_(.*)_")[,2]  
spe$tissue <- str_match(spe$sample, "^.*_.*_(.*)")[,2] 
spe$tissue[spe$tissue=="TU"] = "PT"

meta <- read.csv(file.path(path, '20230314_metadata_complete_comma.csv'))
#meta <- meta[meta$Tissue == 'BM',]

spe$puncture_side <- meta$Puncture_side[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$infiltration_right <- meta$Infiltration_right[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$infiltration_left <- meta$Infiltration_left[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$imc_exclude <- meta$IMC_exclude[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$sca_exclude <- meta$SCA_exclude[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$imc_score <- meta$IMC_score[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$staining_date <- meta$Staining_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$stainer <- meta$Stainer[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$imc_date <- meta$IMC_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$DE_date <- meta$DE_date[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$DE_year <- sub("-.*", "", spe$DE_date)
spe$control <- meta$control[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$study_id <- meta$Study_ID[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$timepoint <- meta$Timepoint[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$stage <- meta$stage4[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$progression <- meta$efs[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod <- meta$tod[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$efs_dur <- meta$efs_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod_date <- meta$tod_dat1[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$tod_dur <- meta$tod_dur[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$rez_date <- meta$rez_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$prog_date <- meta$prog_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$event_date <- meta$event_dat[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$MYCN_amp <- meta$MYCN_amp[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ALK_amp <- meta$ALK_amp[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ALK_mut <- meta$ALK_mut[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$ATRX_del <- meta$ATRX_del[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$TERT_rear <- meta$TERT_rear[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X1p_loss <- meta$X1p_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X1q_gain <- meta$X1q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X11q_loss <- meta$X11q_loss[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]
spe$X17q_gain <- meta$X17q_gain[match(paste(spe$fm_id, spe$tissue), paste(meta$Sample, meta$Tissue))]

spe$ROI <- str_match(spe$sample_id, "^.*_.*_.*_(.*_.*)_.*_.*$")[,2]
spe$year <- gsub(spe$sample_id, pattern="^.*_(.*)-.*", replacement="\\1")
```


**Transform counts**

The distribution of expression counts across cells is often observed to be
skewed towards the right side meaning lots of cells display low counts and few
cells have high counts. To avoid analysis biases from these high-expressing
cells, the expression counts are commonly transformed or clipped.

Here, we perform counts transformation using an inverse hyperbolic sine
function. This transformation is commonly applied to [flow cytometry
data](https://support.cytobank.org/hc/en-us/articles/206148057-About-the-Arcsinh-transform).
The `cofactor` here defines the expression range on which no scaling is
performed. While the `cofactor` for CyTOF data is often set to `5`, IMC data
usually display much lower counts. We therefore apply a `cofactor` of `1`.

However, other transformations such as `log(counts(spe) + 0.01)` should be
tested when analysing IMC data.

```{r transform-counts, message=FALSE}

#assay(spe, "exprs") <- asinh(counts(spe)/1)
#assay(spe, "log") <- log(counts(spe) + 0.1)
#assay(spe, "log10") <- log10(counts(spe) + 0.1)

assay(spe, "exprs") <- counts(spe)
```

Low (negative) values are being are reduced and stretched (not as much as high data), while high values are given more similar y values (stretched), and are increased. Co-factor is used to scale down data before transforming it.

**Check for technical confounders (batch effects)**

```{r plot by sample}
selected_samples <- unique(spe$sample)[10:20]
marker <- '152Sm-CD3_Sm152_mean-80'

library(dittoSeq)
dittoRidgePlot(spe[,spe$sample%in%selected_samples], var = marker, group.by = "sample", assay = "counts") +
    ggtitle(paste(marker,  "- before transformation"))

# dittoRidgePlot(spe[,spe$sample%in%selected_samples], var = marker, group.by = "sample", assay = "exprs", max=3) +
#     ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe[,spe$sample%in%selected_samples], var = marker, group.by = "sample", assay = "log", max=1) +
#     ggtitle(paste(marker,  "- after transformation"))
# 
# dittoRidgePlot(spe[,spe$sample%in%selected_samples], var = marker, group.by = "sample", assay = "log10", max=1) +
#     ggtitle(paste(marker,  "- after transformation"))
```

```{r plot by staining date}

marker <- '152Sm-CD3_Sm152'
selecting_staining_dates <- unique(spe$staining_date)[1:5]
dittoRidgePlot(spe[,spe$staining_date%in%selecting_staining_dates], var = marker, group.by = "staining_date", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe[,spe$staining_date%in%selecting_staining_dates], var = marker, group.by = "staining_date", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))

```


```{r plot by year}

marker <- 'HLA-DR'
dittoRidgePlot(spe, var = marker, group.by = "year", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe, var = marker, group.by = "year", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))

```

```{r plot by RoI}

marker <- 'CD3'
selected_roi <- unique(spe$sample)[3]
dittoRidgePlot(spe[,spe$sample%in%selected_roi], var = marker, group.by = "ROI", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe[,spe$sample%in%selected_roi], var = marker, group.by = "ROI", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))
```

```{r plot by patient}

selected_patients <- unique(spe$study_id)[10:20]
marker <- 'CD45'

dittoRidgePlot(spe[,spe$study_id%in%selected_patients], var = marker, group.by = "study_id", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe[,spe$study_id%in%selected_patients], var = marker, group.by = "study_id", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))

```


```{r plot by stainer}

marker <- 'HLA-DR'
dittoRidgePlot(spe, var = marker, group.by = "stainer", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe, var = marker, group.by = "stainer", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))
```


```{r plot by IMC date}

marker <- 'HLA-DR'
dittoRidgePlot(spe, var = marker, group.by = "imc_date", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe, var = marker, group.by = "imc_date", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))
```


```{r plot by tissue}

marker <- 'Ki-67'

dittoRidgePlot(spe, var = marker, group.by = "tissue", assay = "exprs") +
    ggtitle(paste(marker,  "- after transformation"))

# dittoRidgePlot(spe, var = marker, group.by = "tissue", assay = "log") +
#     ggtitle(paste(marker,  "- after transformation"))

```

**Define interesting channels**

For downstream analysis such as visualization, dimensionality reduction and
clustering, only a subset of markers should be used. As convenience,
we can store an additional entry in the `rowData` slot that specifies the 
markers of interest. Here, we deselect the nuclear markers and keep all other
biological targets.

```{r select-features}
rowData(spe)$use_channel <- !grepl("Histone|H4|Iridium|DAPI|136Ba|131Xe|134Xe|GD2-IF|CD56-IF", rownames(spe))

```

**Define color schemes**

We will already at this point define color schemes for different metadata
entries which will be helpful for downstream data visualizations. We will use
colors from the `RColorBrewer` and `dittoSeq` package but any other coloring
package will suffice.

```{r define-colors, message=FALSE}
# library(RColorBrewer)
# color_vectors <- list()
# sample <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$sample))], 
#                 unique(spe$sample))
# 
# ROI <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$ROI))], 
#                 unique(spe$ROI))
# 
# 
# staining_date <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$staining_date))], 
#                 unique(spe$staining_batch))
# 
# year <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$year))], 
#                 unique(spe$year))
# 
# color_vectors$sample <- sample
# color_vectors$year <- year
# color_vectors$ROI <- ROI
# color_vectors$staining_date <- staining_date
# 
# metadata(spe)$color_vectors <- color_vectors
```

## Read in images {#read-images}

The `cytomapper` package allows multi-channel image handling and visualization
within the Bioconductor framework. The most common data format for multi-channel
images or segmentation masks is the TIFF file format, which is used by `steinbock`
to save images. 

Here, we will read in multi-channel images and segmentation masks into a
[CytoImageList](https://www.bioconductor.org/packages/release/bioc/vignettes/cytomapper/inst/doc/cytomapper.html#5_The_CytoImageList_object)
data container. It allows storing multiple multi-channel images and requires
matched channels across all images within the object.

The `loadImages` function is used to read in processed multi-channel images and
their corresponding segmentation masks as generated by `steinbock`. The
multi-channel images are saved as 32-bit images while the segmentation masks are
saved as 16-bit images. To correctly scale pixel values of the segmentation
masks when reading them in set `as.is = TRUE`.

```{r read-images}
images <- loadImages(file.path(path,"img"))
masks <- loadImages(file.path(path, "masks"), as.is = TRUE)
```

In the case of multi-channel images, it is beneficial to set the `channelNames`
for easy visualization. Using the `steinbock` framework, the channel order of the
single-cell data matches the channel order of the multi-channel images. However,
it is recommended to make sure that the channel order is known.

```{r set-channelNames}
channelNames(images) <- rownames(spe)
```

For visualization shown in Section \@ref(image-visualization) we will need to
add additional metadata to the `elementMetadata` slot of the `CytoImageList`
objects. This slot is easily accessible using the `mcols` function.

Here, we will save the matched `sample_id`, `patient_id` and `indication`
information within the `elementMetadata` slot of the multi-channel images and
segmentation masks objects. It is crucial that the order of the images in 
both `CytoImageList` objects is the same.

```{r add-metadata}
#all.equal(names(images), names(masks))
#sample <- sub("_ROI.*", "", names(images))

#mcols(images) <- mcols(masks) <- DataFrame(image_id = names(images),
#                                           sample = sample)
```

## Save objects

Finally, the generated data objects can be saved for further downstream 
processing and analysis.

```{r save-objects-read-data}

spe <- spe[,!spe$sca_exclude%in%c("yes")]

saveRDS(spe, file.path(path, "data/spe_1px.rds"))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
