==> code/R/IMCData_analysis/02_spillover_correction.Rmd <==
---
title: "Correct spillover"
output: html_document
date: '2022-08-12'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Spillover correction

**Original scripts:** *Vito Zanotelli*, **adapted/maintained by:** *Nils Eling*

This section highlights how to generate a spillover matrix from individually
acquired single metal spots on an agarose slide. Each spot needs to be imaged as
its own acquisition/ROI and individual .txt files containing the pixel
intensities per spot need to be available. For complete details on the spillover
correction approach, please refer to [the original
publication](https://www.cell.com/cell-systems/fulltext/S2405-4712(18)30063-2) [@Chevrier2017].

**Spillover slide preparation:**  

* Prepare 2% agarose in double distilled H$_2$O in a beaker and melt it in a microwave until well dissolved. 
* Dip a blank superfrost plus glass microscope slide into the agarose and submerge it until the label. 
* Remove the slide and prop it up against a support to allow the excess agarose to run off onto paper towels.  
* Allow the slide to dry completely (at least 30 minutes).  
* Retrieve all the antibody conjugates used in the panel for which the spillover matrix is to be generated and place them on ice. 
* Arrange them in a known order (e.g. mass of the conjugated metal).  
* Pipette 0.3 µl spots of 0.4% trypan blue dye into an array on the slide. Prepare one spot per antibody, and make sure the spots are well separated. 
* Pipette 0.3 µl of each antibody conjugate (usually at 0.5 mg/ml) onto a unique blue spot, taking care to avoid different antibodies bleeding into each other. Note the exact location of each conjugate on the slide. 
* Let the spots dry completely, at least 1 hour. 

**Spillover slide acquisition:**  

* Create a .jpeg or .png image of the slide using a mobile phone camera or flat-bed scanner.  
* In the CyTOF software, create a new file and import the slide image into it. 
* Create a panorama across all the spots to visualize their locations.  
* Within each spot, create a region of interest (ROI) with a width of 200 pixels and a height of 10 pixels.  
* Name each ROI with the mass and name of the metal conjugate contained in the spot, e.g "Ir193" or "Ho165". This will be how each .txt file is named. 
* Set the profiling type of each ROI to "Local".  
* Apply the antibody panel to all the ROIs. This panel should contain all (or more) of the isotopes in the panel, with the correct metal specified. For example: if the metal used is Barium 138, make sure this, rather than Lanthanum 138, is selected. 
* Save the file, make sure "Generate Text File" is selected, and start the acquisition. 

This procedure will generate an MCD file similar to the one available on zenodo:
[https://zenodo.org/record/5949116/files/compensation.zip](https://zenodo.org/record/5949116/files/compensation.zip)

The original code of the spillover correction manuscript is available on Github
[here](https://github.com/BodenmillerGroup/cyTOFcompensation); however, due to
changes in the
[CATALYST](https://bioconductor.org/packages/release/bioc/html/CATALYST.html)
package, users were not able to reproduce the analysis using the newest software
versions. The following workflow uses the newest package versions to generate a
spillover matrix.

In brief, the highlighted workflow comprises 5 steps:

1. Reading in the data
2. Quality control 
3. (Optional) pixel binning
4. "Debarcoding" for pixel assignment
5. Pixel selection for spillover matrix estimation
6. Spillover matrix generation
7. Saving the results
8. Single-cell compensation
9. Image compensation

## Generate the spillover matrix

In the first step, we will generate a spillover matrix based on the single-metal
spots and save it for later use.

### Read in the data

Here, we will read in the individual .txt files into a `SingleCellExperiment`
object. This object can be used directly by the `CATALYST` package to estimate
the spillover.

For this to work, the .txt file names need to contain the spotted metal isotope
name. By default, the first occurrence of the isotope in the format `(mt)(mass)`
(e.g. `Sm152` for Samarium isotope with the atomic mass 152) will be used as
spot identifier. Alternatively, a named list of already read-in pixel intensities
can be provided. For more inforation, please refer to the man page `?readSCEfromTXT`.

For further downstream analysis, we will asinh-transform the data using a
cofactor of 5; a common transformation for CyTOF data [@Bendall2011].

```{r read-txts, message=FALSE}
library(imcRtools)

path_to_spillover_data <- "/mnt/Multimodal_Imaging_Daria/20220518_DL_SpilloverCorrection/Extra"

# Create SingleCellExperiment from .txt files
sce <- readSCEfromTXT(path_to_spillover_data) 
assay(sce, "exprs") <- asinh(counts(sce)/5)
assay(sce, "exprs_cof1") <- asinh(counts(sce)/1)
```

### Quality control

In the next step, we will observe the median pixel intensities per spot and
threshold on medians < 200 counts.
These types of visualization serve two purposes:

1. Small median pixel intensities (< 200 counts) might hinder the robust
estimation of the channel spillover. In that case, consecutive pixels can be
summed (see [Optional pixel binning](#pixel_binning)).

2. Each spotted metal (row) should show the highest median pixel intensity in its
corresponding channel (column). If this is not the case, either the naming of the
.txt files was incorrect or the incorrect metal was spotted.

```{r QC-heatmap, message = FALSE, fig.width=7, fig.height=7}
# Log10 median pixel counts per spot and channel
plotSpotHeatmap(sce)

# Thresholded on 200 pixel counts
plotSpotHeatmap(sce, log = FALSE, threshold = 200)

sce1<- sce
```

As we can see, all median pixel intensities are > 200 counts for each spot.
We also observe acquired channels for which no spot was placed (Xe134, Ir191, Ir193). 

### Optional pixel binning {#pixel_binning}

In cases where median pixel intensities are low (< 200 counts), consecutive
pixels can be summed to increase the robustness of the spillover estimation.
The `imcRtools` package provides the `binAcrossPixels` function,
which performs aggregation for each channel across `bin_size` consecutive pixels 
per spotted metal.

```{r binning, message=FALSE, fig.width=7, fig.height=7}
# Define grouping
bin_size = 10

sce2 <- binAcrossPixels(sce, bin_size = bin_size)
assay(sce2, "exprs") <- asinh(counts(sce2)/5)
assay(sce2, "exprs_cof1") <- asinh(counts(sce2)/1)

# Log10 median pixel counts per spot and channel
plotSpotHeatmap(sce2)

# Thresholded on 200 pixel counts
plotSpotHeatmap(sce2, log = FALSE, threshold = 200)

sum(counts(sce2[2,sce2$sample_mass=="145"])) #channel 145 is zero, find out why
```

Here, we can see an increase in the median pixel intensities and accumulation of
off-diagonal signal. Due to already high original pixel intensities, we will
refrain from aggregating across consecutive pixels for this demonstration.

### Filtering incorrectly assigned pixels

The following step uses functions provided by the `CATALYST` package to
"debarcode" the pixels. Based on the intensity distribution of all channel,
pixels are assigned to their corresponding barcode; here this is the already
known metal spot. This procedure serves the purpose to identify pixels that
cannot be robustly assigned to the spotted metal. Pixels of such kind can be
regarded as "noisy", "background" or "artefacts" that should be removed prior to
spillover estimation.

We will also need to specify which channels were spotted (argument `bc_key`).
This information is directly contained in the `colData(sce)` slot.
To facilitate visualization, we will order the `bc_key` by mass.

The general workflow for pixel debarcoding is as follows:

1. assign a preliminary metal mass to each pixel
2. for each pixel, estimate a cutoff parameter for the distance between 
positive and negative pixel sets
3. apply the estimated cutoffs to identify truly positive pixels

```{r debarcoding, message=FALSE}
library(CATALYST)

bc_key <- as.numeric(unique(sce1$sample_mass))
bc_key <- bc_key[order(bc_key)]

sce1 <- assignPrelim(sce1, bc_key = bc_key)
sce1 <- estCutoffs(sce1)
sce1 <- applyCutoffs(sce1)

bc_key <- as.numeric(unique(sce2$sample_mass))
bc_key <- bc_key[order(bc_key)]

sce2 <- assignPrelim(sce2, bc_key = bc_key)
sce2 <- estCutoffs(sce2)
sce2 <- applyCutoffs(sce2)
```

The obtained `SingleCellExperiment` now contains the additional `bc_id` entry.
For each pixel, this vector indicates the assigned mass (e.g. `161`) or
`0`, meaning unassigned. 

This information can be visualized in form of a heatmap:

```{r assignment-heatmap, fig.width=7, fig.height=7}
library(pheatmap)
cur_table <- table(sce1$bc_id, sce1$sample_mass)

pheatmap(log10(cur_table + 1),
         cluster_rows = FALSE,
         cluster_cols = FALSE)

cur_table <- table(sce2$bc_id, sce2$sample_mass)

pheatmap(log10(cur_table + 1),
         cluster_rows = FALSE,
         cluster_cols = FALSE)
```

We can see here, that all pixels were assigned to the right mass and that all
pixel sets are made up if > 1000 pixels.

However, in cases where incorrect assignment occurred or where few pixels were
measured for some spots, the `imcRtools` package exports a simple helper
function to exclude pixels based on these criteria:

```{r pixel-filtering}
sce1 <- filterPixels(sce1, minevents = 40, correct_pixels = TRUE)

sce2 <- filterPixels(sce2, minevents = 40, correct_pixels = TRUE)
```

### Compute spillover matrix

Based on the single-positive pixels, we use the `CATALYST::computeSpillmat()`
function to compute the spillover matrix and `CATALYST::plotSpillmat()` to
visualize it. The `plotSpillmat` function checks the spotted and acquired
metal isotopes against a pre-defined `CATALYST::isotope_list()`. In the demonstration,
the `Ar80` channel was additionally acquired to check for deviations in signal
intensity. `Ar80` needs to be added to a custom `isotope_list` object for
visualization. 

```{r compute-spillover, fig.width=10, fig.height=10}

sce1_all <- computeSpillmat(sce1, interactions="all")
sce2_all <- computeSpillmat(sce2, interactions="all")

sce1 <- computeSpillmat(sce1, interactions="default")
sce2 <- computeSpillmat(sce2, interactions="default")

isotope_list <- CATALYST::isotope_list

plotSpillmat(sce1, isotope_list = isotope_list)
plotSpillmat(sce2, isotope_list = isotope_list)

plotSpillmat(sce1_all, isotope_list = isotope_list)
plotSpillmat(sce2_all, isotope_list = isotope_list)
```
We can see that when we use binning, so summing up consecutive pixels, then the result gets less accurate, e.g. we se no spillover from 167 (H4K12Ac) to 168 (GATA3), which we know we have based on the IMC images of the compensation experiment. The computeSpillMat excludes all interactions that are considered chemically and physically non-sensible, which is why we do not see the spillover from Pt196 to Lu175 anymore. Based on detailed examination, we know that there was an error during the spillover experiment. We therefore decide to use the spillover matrix without binning and default interactions for the compensation. 


As we can see, the largest spillover appears in `In113 --> In115` and we also
observe the `+16` oxide impurities for e.g. `Nd148 --> Dy164`.

Here, we will save the spillover matrix as .csv.

### Save spillover matrix as csv

Next, we will write out the spillover matrix as .csv file for possible later use.

```{r write_sm_csv}

sm1 <- metadata(sce1)$spillover_matrix
sm2 <- metadata(sce2)$spillover_matrix
sm1_all <- metadata(sce1_all)$spillover_matrix
sm2_all <- metadata(sce2_all)$spillover_matrix
write.csv(sm1, file.path(path,"data/spillover_matrix_no-binning_default-interactions.csv"))
write.csv(sm2, file.path(path,"data/spillover_matrix_binning_default-interactions.csv"))
write.csv(sm1_all, file.path(path,"data/spillover_matrix_no-binning_all-interactions.csv"))
write.csv(sm2_all, file.path(path,"data/spillover_matrix_binning_all-interactions.csv"))
```

## Single-cell data compensation

The `CATALYST` package can be used to perform spillover compensation on the
**single-cell mean intensities**. Here, the `SpatialExperiment` object generated
in Section \@ref(read-data) is read in. The `CATALYST` package requires an entry
to `rowData(spe)$channel_name` for the `compCytof` function to run. This entry
should contain the metal isotopes in the form (mt)(mass)Di (e.g., `Sm152Di` for
Samarium isotope with the atomic mass 152). 

The `compCytof` function performs channel spillover compensation on the mean
pixel intensities per channel and cell. Here, we will not overwrite the assays
in the `SpatialExperiment` object to later highlight the effect of compensation.
As shown in Section \@ref(read-data), also the compensated counts are
asinh-transformed using a cofactor of 1.

Non-metal columns present in the SPE (single-cell data), but not in the SCE (spillover matrix) will be added such that they do neither receive nor cause spill, as explained [here](https://rdrr.io/bioc/CATALYST/man/compCytof.html).

```{r single-cell-compensation}
path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20221010_Steinbock_Mesmer_IFbased/BM"

spe <- readRDS(file.path(path,"data/BM_spe.rds"))

rowData(spe)$channel_name <- paste0(rowData(spe)$channel, "Di")

spe <- compCytof(spe, sm, 
                 transform = TRUE, cofactor = 1,
                 isotope_list = isotope_list, 
                 overwrite = FALSE)
```

To check the effect of channel spillover compensation, the expression of markers
that are affected by spillover (e.g., HLA-DR in channel Nd143 and PRPH in
channel Nd144) can be visualized in form of scatter plots before and after
compensation.

```{r visualize-single-cell-spillover, message=FALSE}
library(dittoSeq)
library(patchwork)

before <- dittoScatterPlot(spe, x.var = "HLA-DR", y.var = "PRPH",
                           assay.x = "exprs", assay.y = "exprs") +
    ggtitle("Before compensation")

after <- dittoScatterPlot(spe, x.var = "HLA-DR", y.var = "PRPH",
                          assay.x = "compexprs", assay.y = "compexprs") +
    ggtitle("After compensation")
before + after
```

We observe that the spillover Nd43 --> Nd144 was successfully corrected.
Optionally, the compensated counts can be transferred to the counts slot. We will not 
perform this step, as we want to have the option to choose between non- and compensated counts/exprs.

```{r overwrite-assays}

#assay(spe, "counts") <- assay(spe, "compcounts") 
#assay(spe, "exprs") <- assay(spe, "compexprs") 
#assay(spe, "compcounts") <- assay(spe, "compexprs") <- NULL
```

## Image compensation

The [cytomapper](https://github.com/BodenmillerGroup/cytomapper) package allows channel
spillover compensation directly on **multi-channel images**. 
The `compImage` function takes a `CytoImageList` object and the estimated
spillover matrix as input. More info on how to work with `CytoImageList`
objects can be seen in Section \@ref(image-visualization).

At this point, we can read in the `CytoImageList` object containing multi-channel
images as generated in Section \@ref(read-data).
The `channelNames` need to be set according to their metal isotope in the form
(mt)(mass)Di and therefore match `colnames(sm)`.

```{r read-in-image, message=FALSE}
library(cytomapper)
#images <- readRDS("data/images.rds")
rowData(spe)$channel_name <- paste0(rowData(spe)$channel, "Di")
channelNames(images) <- rowData(spe)$channel_name
```

The CATALYST package provides the `adaptSpillmat` function that corrects the
spillover matrix in a way that rows and columns match a predefined set of
metals. Please refer to `?compCytof` for more information how metals in the
spillover matrix are matched to acquired channels in the `SingleCellExperiment`
object.

The spillover matrix can now be adapted to exclude channels that are not part of
the measurement (`keep == 0`).

```{r}
library(tiff)
panel <- read.csv(file.path(path,"panel_sm.csv"))
adapted_sm <- adaptSpillmat(sm, paste0(panel$channel[panel$keep == 1], "Di"), 
                            isotope_list = isotope_list)
```

The adpated spillover matrix now matches the `channelNames` of the
`CytoImageList` object and can be used to perform pixel-level spillover
compensation.

```{r image-compensation, message = FALSE}
library(BiocParallel)
library(parallel)
images <- getChannels(images, 1:44) #excluding IF channels
images_short <- images[sample(length(images), 20)]
saveRDS(images, file.path(path,"data/images.rds"))
rm(images)
gc()
images_comp <- compImage(images_short, adapted_sm) 
                         BPPARAM = MulticoreParam()
gc()
```

As a sanity check, we will visualize the image before and after compensation:

```{r image-visualization}
# Before compensation
plotPixels(images[5], colour_by = "Yb173Di", 
           image_title = list(text = "Yb173 (Ecad) - before", position = "topleft"), 
           legend = NULL, bcg = list(Yb173Di = c(0, 4, 1)))
plotPixels(images[5], colour_by = "Yb174Di", 
           image_title = list(text = "Yb174 (CD303) - before", position = "topleft"), 
           legend = NULL, bcg = list(Yb174Di = c(0, 4, 1)))
# After compensation
plotPixels(images_comp[5], colour_by = "Yb173Di",
           image_title = list(text = "Yb173 (Ecad) - after", position = "topleft"), 
           legend = NULL, bcg = list(Yb173Di = c(0, 4, 1)))
plotPixels(images_comp[5], colour_by = "Yb174Di", 
           image_title = list(text = "Yb174 (CD303) - after", position = "topleft"),
           legend = NULL, bcg = list(Yb174Di = c(0, 4, 1)))
```

For convenience, we will re-set the `channelNames` to their biological targets:

```{r re-set-channels}
channelNames(images_comp) <- rownames(spe)
```

## Write out compensated images

In the final step, the compensated images are written out as 16-bit TIFF
files:

```{r write-images, message=FALSE, results='hide', error=TRUE}
library(tiff)
dir.create("data/comp_img")
lapply(names(images_comp), function(x){
  writeImage(as.array(images_comp[[x]])/(2^16 - 1), 
             paste0("data/comp_img/", x, ".tiff"),
             bits.per.sample = 16)
})
```


## Save objects

For further downstream analysis, the compensated `SpatialExperiment` are saved:

```{r save-objects-spill}
#saveRDS(spe, file.path(path,"data/spe.rds"))
```


## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
