==> code/R/IMCData_analysis/08_compare_clustering_methods.Rmd <==
---
title: "Compare clustering methods (post-curation)"
output: html_document
date: '2023-05-15'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Cell phenotyping 

A common step during single-cell data analysis is the annotation of cells based
on their phenotype. Defining cell phenotypes is often subjective and relies
on previous biological knowledge. The [Orchestrating Single Cell Analysis with Bioconductor](https://bioconductor.org/books/3.14/OSCA.basic/cell-type-annotation.html) book
presents a number of approaches to phenotype cells detected by single-cell RNA
sequencing based on reference datasets or prior knowledge of geneset importance.

In highly-multiplexed imaging, target proteins or molecules are manually
selected based on the biological question at hand. It narrows down the feature
space and facilitates the manual annotation of clusters to derive cell
phenotypes. We will therefore discuss and compare a number of clustering
approaches to group cells based on their similarity in marker expression in
Section \@ref(clustering).

Unlike single-cell RNA sequencing or CyTOF data, single-cell data derived from
highly-multiplexed imaging data often suffers from "lateral spillover" between
neighboring cells. This spillover caused by imperfect segmentation often hinders
accurate clustering to define specific cell phenotypes in multiplexed imaging
data. Tools have been developed to correct lateral spillover between cells
[@Bai2021] but the approach requires careful selection of the markers to
correct. 

## Load data

We will first read in the previously generated `SpatialExperiment` object and
sample 2000 cells to visualize cluster membership.

```{r read-data, message=FALSE}
library(SpatialExperiment)

path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20230112_DIMR_Ilastik/20230312_sc_data_combined"

spe <- readRDS(file.path(path,"data/spe_1px_wo-ghosts.rds"))
rowData(spe)$use_channel <- !grepl("CD274|IDO|DAPI|2IF_GD2|3IF_CD56|DNA1|H4K12Ac|H3K9Ac|CXCR2|PNMT|Fibro|mean-80", rownames(spe))

non_zeros <- as.logical(colSums(assay(spe, "exprs")[rowData(spe)$use_channel,])!=0)

spe <- spe[,non_zeros]
set.seed(220619)

```

## Clustering approaches {#clustering}

In the first section to identify cellular phenotypes in the dataset, we will
present clustering approaches that group cells based on their similarity in
marker expression. A number of approaches have been developed to cluster data
derived from single-cell RNA sequencing technologies [@Yu2022] or CyTOF
[@Weber2016]. For demonstration purposes, we will highlight common clustering
approaches that are available in R and have been used for clustering cells
obtained from IMC. Two approaches rely on graph-based clustering and one
approach uses self organizing maps (SOM).

### Rphenograph

The PhenoGraph clustering approach was first described to group cells of a CyTOF
dataset [@Levine2015]. The algorithm first constructs a graph by detecting the
`k` nearest neighbours based on euclidean distance in expression space. In the
next step, edges between nodes (cells) are weighted by their overlap in nearest
neighbor sets. To quantify the overlap in shared nearest neighbor sets, the
jaccard index is used. The Louvain modularity optimization approach is used to
detect connected communities and partition the graph into clusters of cells.
This clustering strategy was used by Jackson, Fischer _et al._  and Schulz _et
al._ to cluster IMC data [@Jackson2020; @Schulz2018].

There are several different PhenoGraph implementations available in R. Here, we
use the one available at
[https://github.com/i-cyto/Rphenograph](https://github.com/i-cyto/Rphenograph).
For large datasets,
[https://github.com/stuchly/Rphenoannoy](https://github.com/stuchly/Rphenoannoy)
offers a more performant implementation of the algorithm.

In the following code chunk, we select the asinh-transformed mean pixel
intensities per cell and channel and subset the channels to the ones containing
biological variation. This matrix is transposed to store cells in rows. Within
the `Rphenograph` function, we select the 45 nearest neighbors for graph
building and louvain community detection (default). The function returns a list
of length 2, the first entry being the graph and the second entry containing the
community object. Calling `membership` on the community object will return
cluster IDs for each cell. These cluster IDs are then stored within the
`colData` of the `SpatialExperiment` object. Cluster IDs are mapped on top of
the UMAP embedding and single-cell marker expression within each cluster are
visualized in form of a heatmap.

It is recommended to test different inputs to `k` as shown in the next section.
Selecting larger values for `k` results in larger clusters.

```{r, echo=FALSE}
start_time <- Sys.time()
```

Here we will test [fastPG](https://www.biorxiv.org/content/10.1101/2020.06.19.159749v1.full.pdf) as the fast implementation of Phenograph, as it is
reported to be more accurate than Rphenograph

```{r rphenograph-1, message=FALSE, eval=FALSE}
library(FastPG)
library(igraph)
library(dittoSeq)
library(viridis)
library(parallel)
library(scater)
library(BiocParallel)
library(RColorBrewer)
library(paletteer)

set.seed(220225)
spe <- runUMAP(spe, subset_row = rowData(spe)$use_channel, exprs_values = "exprs", BPPARAM = MulticoreParam()) 

n_threads <- detectCores()

mat <- t(assay(spe, "exprs")[rowData(spe)$use_channel,])

out <- FastPG::fastCluster(data=mat, k=45, num_threads=n_threads)

clusters <- factor(out$communities)

spe$pg_clusters <- clusters

set.seed(20230317)

n <- length(unique(clusters))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters <- setNames(sample(col_vector_433, n), unique(clusters))

dittoDimPlot(spe, var = "pg_clusters", 
             reduction.use = "UMAP", size = 0.2,
             do.label = TRUE,
             color.panel=col_clusters) +
    ggtitle("Phenograph clusters expression on UMAP")
```


```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}
library(ComplexHeatmap)
library(circlize)

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

set.seed(20230520)
#mat <- t(mat)
tissue <- spe$tissue

#Get same number of PT and BM cells
pt_index <- which(tissue=="PT")
bm_index <- which(tissue=="BM")
rand_cells_pt <- sample(seq_len(length(pt_index)), 4000)
rand_cells_bm <- sample(seq_len(length(bm_index)), 4000)
pt_index_rand <- pt_index[rand_cells_pt]
bm_index_rand <- bm_index[rand_cells_bm]
rand_cells <- c(pt_index_rand, bm_index_rand)



column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue 
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with phenograph clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```


```{r save-spe, message=FALSE}
out_path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20230112_DIMR_Ilastik/20230411_sc_data_combined"
saveRDS(spe, file.path(out_path, "data", "spe_1px_clusters.rds"))
```


```{r, echo=FALSE}
end_time <- Sys.time()
```

The `Rphenograph` function call took 
`r round(as.numeric(difftime(end_time, start_time, units = "mins")), digits = 2)` minutes.



### Shared nearest neighbour graph {#snn-graph}

The [bluster](https://www.bioconductor.org/packages/release/bioc/html/bluster.html) 
package provides a simple interface to cluster cells using a number of different
[clustering approaches](https://www.bioconductor.org/packages/release/bioc/vignettes/bluster/inst/doc/clusterRows.html) and different metrics to [access cluster stability](https://www.bioconductor.org/packages/release/bioc/vignettes/bluster/inst/doc/diagnostics.html).  

For simplicity, we will focus on graph based clustering as this is the most
popular and a fast method for single-cell clustering. The `bluster` package
provides functionalities to build k-nearest neighbour (KNN) and it's weighted
version (shared nearest neighbor; SNN) graphs where nodes represent cells.
The user can chose the number of neighbors to consider (parameter `k`), 
the edge weighting method (parameter `type`) and the community detection
function to use (parameter `cluster.fun`). As all parameters affect the clustering
results, the `bluster` package provides the `clusterSweep` function to test
a number of parameter settings in parallel. In the following code chunk,
we select the markers
of interest. The resulting matrix is transposed to fit to the requirements of 
the bluster package (cells in rows).

We test two different settings for `k`, two for `type` and fix the `cluster.fun`
to `louvain` as this is the most common approach for community detection.
This function call is parallelized by setting the `BPPARAM` parameter. We use
the `approxSilhouette` function to compute the silhouette width for each cell
and compute the average across all cells per parameter setting. 
Please see `?silhouette` for more information on how the silhouette 
width id computed for each cell. A large average silhouette width indicates 
cells that are well clustered.

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r parameter-sweep, message=FALSE, fig.height=7}
library(bluster)
library(BiocParallel)
library(ggplot2)

mat <- t(assay(spe, "exprs")[rowData(spe)$use_channel,])

combinations <- clusterSweep(mat, 
                             BLUSPARAM=SNNGraphParam(),
                             k=c(10L, 20L), 
                             type = c("rank", "jaccard"), 
                             cluster.fun = "louvain",
                             BPPARAM = MulticoreParam(RNGseed = 220427))

sil <- vapply(as.list(combinations$clusters), 
              function(x) mean(approxSilhouette(mat, x)$width), 
              0)

ggplot(data.frame(method = names(sil),
                  sil = sil)) +
    geom_point(aes(reorder(method, sil), sil)) +
    theme_classic(base_size = 15) +
    theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
    xlab("Cluster parameter combination") +
    ylab("Average silhouette width")
```

```{r, echo=FALSE}
end_time <- Sys.time()
```

The cluster parameter sweep took 
`r round(as.numeric(difftime(end_time, start_time, units = "mins")), digits = 2)` minutes.

Performing a cluster sweep takes some time as multiple function calls are run in parallel.
We do however recommend testing a number of different parameter settings to
access clustering performance.

Once parameter settings are known, we can either use the `clusterRows` function
of the `bluster` package to cluster cells or its convenient wrapper function
exported by the
[scran](https://bioconductor.org/packages/release/bioc/html/scran.html) package.
The `scran::clusterCells` function accepts a `SpatialExperiment` (or
`SingleCellExperiment`) object which stores cells in columns. By default, the
function detects the 10 nearest neighbours for each cell, performs rank-based
weighting of edges (see `?makeSNNGraph` for more information) and uses the
`cluster_walktrap` function to detect communities in the graph.

As we can see above, the clustering approach in this dataset with `k` being 
20 and rank-based edge weighting leads to the highest silouhette width.

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r snn-1, message=FALSE}
library(scran)

clusters <- clusterCells(spe[rowData(spe)$use_channel,], 
                         assay.type = "exprs", 
                         BLUSPARAM = SNNGraphParam(k=20, 
                                                  cluster.fun = "louvain",
                                                  type = "rank",
                                                  BPPARAM = MulticoreParam(RNGseed = 220427)))

spe$nn_clusters <- clusters

set.seed(20230317)

n <- length(unique(clusters))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters <- setNames(sample(col_vector_433, n), unique(clusters))

dittoDimPlot(spe, var = "nn_clusters", 
             reduction.use = "UMAP", size = 0.2,
             do.label = TRUE,
             color.panel=col_clusters) +
    ggtitle("SNN clusters expression on UMAP")
```
```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}
library(ComplexHeatmap)
library(circlize)

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

mat <- t(assay(spe, "exprs")[rowData(spe)$use_channel,])
mat <- t(mat)
tissue <- spe$tissue

#Get same number of PT and BM cells
pt_index <- which(tissue=="PT")
bm_index <- which(tissue=="BM")
rand_cells_pt <- sample(seq_len(length(pt_index)), 4000)
rand_cells_bm <- sample(seq_len(length(bm_index)), 4000)
pt_index_rand <- pt_index[rand_cells_pt]
bm_index_rand <- bm_index[rand_cells_bm]
rand_cells <- c(pt_index_rand, bm_index_rand)



column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue 
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with SNN clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```



```{r, echo=FALSE}
end_time <- Sys.time()
```

```{r save-spe, message=FALSE}
out_path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20230112_DIMR_Ilastik/20230411_sc_data_combined"
saveRDS(spe, file.path(out_path, "data", "spe_1px_clusters.rds"))
```

The shared nearest neighbor graph clustering approach took 
`r round(as.numeric(difftime(end_time, start_time, units = "mins")), digits = 2)` minutes.

This function was used by [@Tietscher2022] to cluster cells obtained by IMC. Setting
`type = "jaccard"` performs clustering similar to `Rphenograph` above and [Seurat](https://satijalab.org/seurat/articles/pbmc3k_tutorial.html#cluster-the-cells-1).


### Self organizing maps

An alternative to graph-based clustering is offered by the
[CATALYST](https://bioconductor.org/packages/release/bioc/html/CATALYST.html)
package. The `cluster` function internally uses the
[FlowSOM](https://bioconductor.org/packages/release/bioc/html/FlowSOM.html)
package to group cells into 100 (default) clusters based on self organizing maps
(SOM). In the next step, the
[ConsensusClusterPlus](https://bioconductor.org/packages/release/bioc/html/ConsensusClusterPlus.html)
package is used to perform hierarchical consensus clustering of the previously
detected 100 SOM nodes into 1 to `maxK` clusters. Cluster stability for each `k`
can be assessed by plotting the `delta_area(spe)`. The optimal number
of clusters can be found by selecting the `k` at which a plateau is reached.
In the example below, the optimal `k` lies somewhere around 13.

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r flowSOM-1, message=FALSE}
library(CATALYST)

# Run FlowSOM and ConsensusClusterPlus clustering
spe <- cluster(spe, 
               features = rownames(spe)[rowData(spe)$use_channel],
               maxK = 30,
               seed = 220410)

# Assess cluster stability
delta_area(spe)

spe$som_clusters <- cluster_ids(spe, "meta25")

set.seed(20230317)

n <- length(unique(clusters))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]
col_clusters <- setNames(sample(col_vector_433, n), unique(clusters))


dittoDimPlot(spe, var = "som_clusters", 
             reduction.use = "UMAP", size = 0.2,
             do.label = TRUE,
             color.panel=col_clusters) +
    ggtitle("SOM clusters expression on UMAP")
```

```{r single-cell-heatmap-clusters, message=FALSE, fig.height=17}
library(ComplexHeatmap)
library(circlize)

col_marker = colorRamp2(c(0, 2.5, 0.5), c("blue", "white", "red"))
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

mat <- t(assay(spe, "exprs")[rowData(spe)$use_channel,])
mat <- t(mat)
tissue <- spe$tissue

#Get same number of PT and BM cells
pt_index <- which(tissue=="PT")
bm_index <- which(tissue=="BM")
rand_cells_pt <- sample(seq_len(length(pt_index)), 4000)
rand_cells_bm <- sample(seq_len(length(bm_index)), 4000)
pt_index_rand <- pt_index[rand_cells_pt]
bm_index_rand <- bm_index[rand_cells_bm]
rand_cells <- c(pt_index_rand, bm_index_rand)



column_ha = HeatmapAnnotation(cluster = clusters[rand_cells],
                              tissue = tissue[rand_cells],
                              col=list(cluster=col_clusters, 
                                       tissue=col_tissue 
                                       ), 
                              show_legend=T)


Heatmap(mat[,rand_cells], 
        column_title = "Single-cell heatmap with SOM clusters (1-px mean)",
        top_annotation = column_ha, 
        cluster_columns=F,
        column_order= order(clusters[rand_cells], decreasing=F),
        show_column_dend = F,
        show_column_names=F,
        col=viridis(100)
        )
```



```{r, echo=FALSE}
end_time <- Sys.time()
```

```{r save-spe, message=FALSE}
out_path <- "/mnt/Multimodal_Imaging_Daria/_Data_Analysis/_tmp_daria/Image_analysis/20230112_DIMR_Ilastik/20230411_sc_data_combined"
saveRDS(spe, file.path(out_path, "data", "spe_1px_clusters.rds"))
```

Running FlowSOM clustering took `r round(as.numeric(difftime(end_time, start_time, units = "mins")), digits = 2)` minutes.

The `FlowSOM` clustering approach has been used by [@Hoch2022] to sub-cluster tumor
cells as measured by IMC.

### Compare between clustering approaches

Finally, we can compare the results of different clustering approaches. For
this, we visualize the number of cells that are shared between different
clustering results in a pairwise fashion. In the following heatmaps a high match
between clustering results can be seen for those clusters that are uniquely
detected in both approaches.

First, we will visualize the match between the three different approaches.

```{r raw-counts, message=FALSE, fig.height=10}
library(patchwork)
library(pheatmap)
library(gridExtra)

tab1 <- table(paste("Rphenograph", spe$pg_clusters), 
              paste("SNN", spe$nn_clusters))
tab2 <- table(paste("Rphenograph", spe$pg_clusters), 
              paste("SOM", spe$som_clusters))
tab3 <- table(paste("SNN", spe$nn_clusters), 
              paste("SOM", spe$som_clusters))

pheatmap(log10(tab1 + 10), color = viridis(100))
pheatmap(log10(tab2 + 10), color = viridis(100))
pheatmap(log10(tab3 + 10), color = viridis(100))
```

We observe that `Rphenograph` and the shared nearest neighbor (SNN) approach by
`scran` show similar results (first heatmap above). For example, Rphenograph
cluster 18 (a tumor cluster) is perfectly captured by SNN cluster 14. On the
other hand, the Neutrophil cluster (SNN cluster 6) is split into Rphenograph
cluster 5 and Rphenograph cluster 9. A common approach
is to now merge clusters that contain similar cell types and annotate them by
hand.


### Further clustering notes {#clustering-notes}

The `bluster` package provides a number of metrics to assess cluster stability
[here](https://www.bioconductor.org/packages/release/bioc/vignettes/bluster/inst/doc/diagnostics.html).
For brevity we only highlighted the use of the silhouette width but different
metrics should be tested to assess cluster stability.

To assign cell types to clusters, we manually annotate clusters based on their
marker expression. For example, SNN cluster 11 (clustering of the integrated
cells) shows high, homogeneous expression of CD20 and we might therefore label
this cluster as B cells. The next chapter \@ref(single-cell-visualization) will 
highlight single-cell visualization methods that can be helpful for manual 
cluster annotations. 

An example how to label clusters can be seen below:

```{r, message=FALSE}
library(dplyr)
cluster_celltype <- recode(spe$nn_clusters_corrected,
                            "1" = "Tumor_proliferating",
                            "2" = "Neutrophil",
                            "3" = "Tumor",
                            "4" = "Tumor",
                            "5" = "Stroma",
                            "6" = "Proliferating",
                            "7" = "Myeloid",
                            "8" = "CD8",
                            "9" = "Plasma_cell",
                            "10" = "CD4",
                            "11" = "Bcell",
                            "12" = "Stroma",
                            "13" = "unknown")

spe$cluster_celltype <- cluster_celltype
```

## Classification approach {#classification}

In this section, we will highlight a cell type classification approach based
on ground truth labeling and random forest classification. The rational for
this supervised cell phenotyping approach is to use the information contained
in the pre-defined markers to detect cells of interest. This approach was
used by Hoch _et al._ to classify cell types in a metastatic melanoma IMC
dataset [@Hoch2022].

The antibody panel used in the example data set mainly focuses on immune cell
types and little on tumor cell phenotypes. Therefore we will label the following
cell types:

* Tumor (E-cadherin positive)
* Stroma (SMA, PDGFRb positive)
* Plasma cells (CD38 positive)
* Neutrophil (MPO, CD15 positive)
* Myeloid cells (HLADR positive)
* B cells (CD20 positive)
* B next to T cells (CD20, CD3 positive)
* Regulatory T cells (FOXP3 positive)
* CD8+ T cells (CD3, CD8 positive)
* CD4+ T cells (CD3, CD4 positive)

The "B next to T cell" phenotype (`BnTcell`) is commonly observed in immune
infiltrated regions measured by IMC. We include this phenotype to account for B
cell/T cell interactions where precise classification into B cells or T cells is
not possible. The exact gating scheme can be seen at
[img/Gating_scheme.pdf](img/Gating_scheme.pdf).

As related approaches, [Astir](https://github.com/camlab-bioml/astir) and
[Garnett](https://cole-trapnell-lab.github.io/garnett/) use pre-defined panel
information to classify cell phenotypes based on their marker expression.

### Manual labeling of cells

The [cytomapper](https://www.bioconductor.org/packages/release/bioc/html/cytomapper.html)
package provides the `cytomapperShiny` function that allows gating of cells
based on their marker expression and visualization of selected cells directly
on the images. 

```{r cytomapperShiny, message=FALSE}
library(cytomapper)
if (interactive()) {
    
    images <- readRDS("data/images.rds")
    masks <- readRDS("data/masks.rds")
    
    cytomapperShiny(object = spe, mask = masks, image = images, 
                    cell_id = "ObjectNumber", img_id = "sample_id")
}
```

The labeled cells for this data set can be accessed at
[zenodo.org/record/7432486](https://zenodo.org/record/7432486) and were
downloaded in Section \@ref(prerequisites). Per image, the `cytomapperShiny`
function allows the export of gated cells in form of a `SingleCellExperiment` or
`SpatialExperiment` object. The cell label is stored in
`colData(object)$cytomapper_CellLabel` and the gates are stored in
`metadata(object)`. In the next section, we will read in and consolidate the
labeled data.

### Define color vectors

For consistent visualization of cell types, we will now pre-define their colors:

```{r cell-type-colors}
celltype <- setNames(c("#3F1B03", "#F4AD31", "#894F36", "#1C750C", "#EF8ECC", 
                       "#6471E2", "#4DB23B", "grey", "#F4800C", "#BF0A3D", "#066970"),
                     c("Tumor", "Stroma", "Myeloid", "CD8", "Plasma_cell", 
                       "Treg", "CD4", "undefined", "BnTcell", "Bcell", "Neutrophil"))

metadata(spe)$color_vectors$celltype <- celltype
```

### Read in and consolidate labeled data

Here, we will read in the individual `SpatialExperiment` objects containing the
labeled cells and concatenate them. In the process of concatenating the 
`SpatialExperiment` objects along their columns, the `sample_id` entry is 
appended by `.1, .2, .3, ...` due to replicated entries. 

```{r, read-in-labeled-data, message=FALSE}
library(SingleCellExperiment)
label_files <- list.files("data/gated_cells", 
                          full.names = TRUE, pattern = ".rds$")

# Read in SPE objects
spes <- lapply(label_files, readRDS)

# Merge SPE objects
concat_spe <- do.call("cbind", spes)
```

In the following code chunk we will identify cells that were labeled multiple
times. This occurs when different cell phenotypes are gated per image and can
affect immune cells that are located inside the tumor compartment.

We will first identify those cells that were uniquely labeled. In the next step,
we will identify those cells that were labeled twice AND were labeled as Tumor
cells. These cells will be assigned their immune cell label. Finally, we will
save the unique labels within the original `SpatialExperiment` object.

For these tasks, we will define a filter function:

```{r}
filter_labels <- function(object, 
                          label = "cytomapper_CellLabel") {
    cur_tab <- unclass(table(colnames(object), object[[label]]))
    
    cur_labels <- colnames(cur_tab)[apply(cur_tab, 1, which.max)]
    names(cur_labels) <- rownames(cur_tab)
    
    cur_labels <- cur_labels[rowSums(cur_tab) == 1]
    
    return(cur_labels)
}
```

This function is now applied to all cells and then only non-tumor cells.

```{r remove-doublets}
labels <- filter_labels(concat_spe)

cur_spe <- concat_spe[,concat_spe$cytomapper_CellLabel != "Tumor"]

non_tumor_labels <- filter_labels(cur_spe)

additional_cells <- setdiff(names(non_tumor_labels), names(labels))

final_labels <- c(labels, non_tumor_labels[additional_cells])

# Transfer labels to SPE object
spe_labels <- rep("unlabeled", ncol(spe))
names(spe_labels) <- colnames(spe)
spe_labels[names(final_labels)] <- final_labels
spe$cell_labels <- spe_labels

# Number of cells labeled per patient
table(spe$cell_labels, spe$patient_id)
```

Based on these labels, we can now train a random forest classifier to classify
all remaining, unlabeled cells.

### Train classifier

In this section, we will use the
[caret](https://topepo.github.io/caret/index.html) framework for machine
learning in R. This package provides an interface to train a number of
regression and classification models in a coherent fashion. We use a random
forest classifier due to low number of parameters, high speed and an observed
high performance for cell type classification [@Hoch2022].

In the following section, we will first split the `SpatialExperiment` object
into labeled and unlabeled cells. Based on the labeled cells, we split
the data into a train (75% of the data) and test (25% of the data) dataset. 
We currently do not provide an independently labeled validation dataset.

The `caret` package provides the `trainControl` function, which specifies model
training parameters and the `train` function, which performs the actual model
training. While training the model, we also want to estimate the best model
parameters. In the case of the chosen random forest model (`method = "rf"`), we
only need to estimate a single parameters (`mtry`) which corresponds to the
number of variables randomly sampled as candidates at each split. To estimate
the best parameter, we will perform a 5-fold cross validation (set within
`trainControl`) over a tune length of 5 entries to `mtry`.

```{r, echo=FALSE}
start_time <- Sys.time()
```

```{r train-classifier, message=FALSE}
library(caret)

# Split between labeled and unlabeled cells
lab_spe <- spe[,spe$cell_labels != "unlabeled"]
unlab_spe <- spe[,spe$cell_labels == "unlabeled"]

# Randomly split into train and test data
set.seed(221029)
trainIndex <- createDataPartition(factor(lab_spe$cell_labels), p = 0.75)
train_spe <- lab_spe[,trainIndex$Resample1]
test_spe <- lab_spe[,-trainIndex$Resample1]

# Specify train parameters for 5-fold cross validation
fitControl <- trainControl(method = "cv",
                           number = 5)

# Select the data for training
cur_mat <- t(assay(train_spe, "exprs")[rowData(train_spe)$use_channel,])

# Train a random forest model for predicting cell labels
# This call also performs parameter tuning
rffit <- train(x = cur_mat, 
               y = factor(train_spe$cell_labels),
               method = "rf", ntree = 1000,
               tuneLength = 5,
               trControl = fitControl)

rffit
```

```{r, echo=FALSE}
end_time <- Sys.time()
```

Training the classifier took 
`r round(as.numeric(difftime(end_time, start_time, units = "mins")), digits = 2)` minutes.

### Classifier performance

We next observe the accuracy of the classifer when predicting cell 
phenotypes across the cross-validation as well as the test dataset.

First, we can visualize the classification accuracy during parameter
tuning:

```{r accuracy-tuning}
ggplot(rffit) + 
  geom_errorbar(data = rffit$results,
                aes(ymin = Accuracy - AccuracySD,
                    ymax = Accuracy + AccuracySD),
                width = 0.4) +
    theme_classic(base_size = 15)
```

The best value for `mtry` is 28 and is used when predicting new data.

It is often recommended to visualize the variable importance of the
classifier. The following plot specifies which variables (markers) are 
most important for classifying the data.

```{r variable-importance, fig.height = 10}
plot(varImp(rffit))
```

As expected, the markers that were used for gating (Ecad, CD3, CD20, HLADR,
CD8a, CD38, FOXP3) were important for classification.

To assess the accuracy, sensitivity, specificity, among other quality measures of
the classifier, we will now predict cell phenotypes in the test data.

```{r model-testing, message=FALSE}
# Select test data
cur_mat <- t(assay(test_spe, "exprs")[rowData(test_spe)$use_channel,])

# Predict cell phenotypes in test data
cur_pred <- predict(rffit, 
                    newdata = cur_mat)
```

While the overall classification accuracy can appear high, we also want
to check if each cell phenotype class is correctly predicted.
For this, we will calculate the confusion matrix between predicted and actual 
cell labels. This measure highlights individual cell phenotype classes that 
were not correctly predicted by the classifier. When setting `mode = "everything"`,
the `confusionMatrix` function returns all available prediction measures including
sensitivity, specificity, precision, recall and the F1 score per cell 
phenotype class.

```{r confusion-matrix}
cm <- confusionMatrix(data = cur_pred, 
                      reference = factor(test_spe$cell_labels), 
                      mode = "everything")

cm
```

To easily visualize these results, we can now plot the true positive rate (sensitivity)
versus the false positive rate (1 - specificity):

```{r specificity-sensitivity, message=FALSE}
library(tidyverse)

data.frame(cm$byClass) %>%
  mutate(class = sub("Class: ", "", rownames(cm$byClass))) %>%
  ggplot() + 
  geom_point(aes(1 - Specificity, Sensitivity, 
                 size = Detection.Rate,
                 fill = class),
             shape = 21) + 
  scale_fill_manual(values = metadata(spe)$color_vectors$celltype) +
  theme_classic(base_size = 15) + 
  ylab("Sensitivity (TPR)") +
  xlab("1 - Specificity (FPR)")
```

We observe high sensitivity and specificity for most cell types. Plasma cells
show the lowest true positive rate with 88% being sufficiently high. The size of
the circle specifies the number of cells per class.

Finally, to observe which cell phenotypes were wrongly classified, we can visualize
the distribution of classification probabilities per cell phenotype class:

```{r prediciton-probability, fig.height=15}
cur_pred <- predict(rffit, 
                    newdata = cur_mat, 
                    type = "prob")
cur_pred$truth <- factor(test_spe$cell_labels)

cur_pred %>%
  pivot_longer(cols = Bcell:Tumor) %>%
  ggplot() +
  geom_boxplot(aes(x = name, y = value, fill = name), outlier.size = 0.5) +
  facet_wrap(. ~ truth, ncol = 1) + 
  scale_fill_manual(values = metadata(spe)$color_vectors$celltype)  +
  theme(panel.background = element_blank(), 
        axis.text.x = element_text(angle = 45, hjust = 1))
```

The boxplots indicate the classification probabilities per class. The classifier
is well trained if classification probabilities are only high for the one 
specific class.

### Classification of new data

In the final section, we will now use the tuned and tested random forest 
classifier to predict the cell phenotypes of the unlabeled data.

First, we predict the cell phenotypes and extract their classification
probabilities.

```{r, unlabeled-prediction}
# Select unlabeled data
cur_mat <- t(assay(unlab_spe, "exprs")[rowData(unlab_spe)$use_channel,])

# Predict cell phenotypes
cell_class <- as.character(predict.train(rffit, 
                                         newdata = cur_mat, 
                                         type = "raw"))
names(cell_class) <- rownames(cur_mat)

table(cell_class)

# Extract classification probabilities
cell_prob <- predict.train(rffit, 
                           newdata = cur_mat, 
                           type = "prob")
```

Each cell is assigned to the class with highest probability. There are however
cases, where the highest probability is low meaning the cell can not be uniquely
assigned to a class. We next want to identify these cells and label them as 
"undefined". Here, we select a maximum classification probability threshold 
of 40% but this threshold needs to be adjusted for other datasets. The adjusted
cell labels are then stored in the `SpatialExperiment` object.

```{r, undefined-cells}
library(ggridges)

# Distribution of maximum probabilities
tibble(max_prob = rowMax(as.matrix(cell_prob)),
       type = cell_class) %>%
    ggplot() +
        geom_density_ridges(aes(x = max_prob, y = cell_class, fill = cell_class)) +
        scale_fill_manual(values = metadata(spe)$color_vectors$celltype) +
        theme_classic(base_size = 15) +
        xlab("Maximum probability") +
        ylab("Cell type") + 
        xlim(c(0,1.2))

# Label undefined cells
cell_class[rowMax(as.matrix(cell_prob)) < 0.4] <- "undefined"

# Store labels in SpatialExperiment onject
cell_labels <- spe$cell_labels
cell_labels[colnames(unlab_spe)] <- cell_class
spe$celltype <- cell_labels 

table(spe$celltype, spe$patient_id)
```

We can now compare the cell labels derived by classification to the different
clustering strategies. The first comparison is against the clustering results
using the asinh-transformed counts.

```{r compare-raw, message=FALSE}
tab1 <- table(spe$celltype, 
              paste("Rphenograph", spe$pg_clusters))
tab2 <- table(spe$celltype, 
              paste("SNN", spe$nn_clusters))
tab3 <- table(spe$celltype, 
              paste("SOM", spe$som_clusters))

pheatmap(log10(tab1 + 10), color = viridis(100))
pheatmap(log10(tab2 + 10), color = viridis(100))
pheatmap(log10(tab3 + 10), color = viridis(100))
```

We can see that Tumor and Myeloid cells span multiple clusters while
Neutrophiles are detected as an individual cluster by all clustering approaches.

We next compare the cell classification against clustering results using the 
integrated cells.

```{r compare-corrected, message=FALSE}
tab1 <- table(spe$celltype, 
              paste("Rphenograph", spe$pg_clusters_corrected))
tab2 <- table(spe$celltype, 
              paste("SNN", spe$nn_clusters_corrected))
tab3 <- table(spe$celltype, 
              paste("SOM", spe$som_clusters_corrected))

pheatmap(log10(tab1 + 10), color = viridis(100))
pheatmap(log10(tab2 + 10), color = viridis(100))
pheatmap(log10(tab3 + 10), color = viridis(100))
```

We observe a high agreement between the shared nearest neighbor clustering
approach using the integrated cells and the cell phenotypes derived by
classification.

In the next sections, we will highlight visualization strategies to verify the
correctness of the phenotyping approach. Specifically, Section
\@ref(outline-cells) shows how to outline identified cell phenotypes on
composite images.

```{r display-images, echo=FALSE, message = FALSE, results='hide'}
# This section serves quality control purposes.
library(testthat)
test_that("images are written out", {
    skip_on_ci()
    expect_true(TRUE)
    cell_types <- unique(spe$celltype)

    markers <- vector(mode = "list", length = length(cell_types))
    names(markers) <- cell_types

    markers[["Bcell"]] <- c("CD20", "CD3")
    markers[["Plasma_cell"]] <- c("CD38")
    markers[["CD4"]] <- c("CD4", "CD8a", "FOXP3")
    markers[["BnTcell"]] <- c("CD20","CD3")
    markers[["CD8"]] <- c("CD4", "CD8a", "FOXP3")
    markers[["Treg"]] <- c("CD4", "CD8a", "FOXP3")
    markers[["Myeloid"]] <- c("CD11c", "CD68", "HLADR", "CD163")
    markers[["Neutrophil"]] <- c("CD15", "MPO")
    markers[["Stroma"]] <- c("SMA", "PDGFRb")
    markers[["Tumor"]] <- c("Ecad")
    markers[["undefined"]] <- c("Ecad", "CD3", "CD20", "MPO", "CD38")
    
    images <- readRDS("data/images.rds")
    masks <- readRDS("data/masks.rds")
    
    images <- cytomapper::normalize(images, separateImages = TRUE)
    images <- cytomapper::normalize(images, separateImages = TRUE, 
                        inputRange = c(0, 0.2))
    
    if (!dir.exists("data/CellTypeValidation")) {
        dir.create("data/CellTypeValidation")
    }
    
    for(j in unique(spe$celltype)){
        cur_sce <- spe[,spe$celltype == j]
        cur_markers <- markers[[j]]
  
        if(dim(cur_sce)[2] == 0) {
            next(j)
        }
    
        if (length(cur_markers) == 1) {
            cur_col <- "red"
            names(cur_col) <- j
            cur_col <- list(cur_col)
            names(cur_col) <- "celltype"
            plotPixels(image = images, 
                object = cur_sce, 
                mask = masks, 
                cell_id = "ObjectNumber",
                img_id = "sample_id", 
                colour_by = cur_markers, 
                outline_by = "celltype",
                image_title = list(text = names(images),
                                cex = 1), 
                colour = cur_col,
                save_plot = list(filename = paste0("data/CellTypeValidation/",
                            j, ".png")))
        } else {
            plotPixels(image = images,
                   object = cur_sce, 
                   mask = masks, 
                   cell_id = "ObjectNumber",
                   img_id = "sample_id", 
                   colour_by = cur_markers, 
                   outline_by = "celltype",
                   image_title = list(text = names(images),
                                  cex = 1),
                   save_plot = list(filename = paste0("data/CellTypeValidation/",
                            j, ".png")))
        }
    }
    expect_true(TRUE)
    
})
```

Finally, we save the updated `SpatialExperiment` object.

```{r phenotyping-save-object}
saveRDS(spe, "data/spe.rds")
```

```{r testing, include=FALSE}
library(testthat)

expect_equal(table(spe$pg_clusters), 
             structure(c(`1` = 3063L, `2` = 2145L, `3` = 1502L, `4` = 2953L, 
`5` = 2185L, `6` = 2973L, `7` = 1366L, `8` = 1508L, `9` = 1769L, 
`10` = 572L, `11` = 1778L, `12` = 4123L, `13` = 3251L, `14` = 4003L, 
`15` = 3880L, `16` = 2387L, `17` = 686L, `18` = 2030L, `19` = 1709L, 
`20` = 702L, `21` = 1159L, `22` = 173L, `23` = 817L, `24` = 1060L
), dim = 24L, dimnames = structure(list(c("1", "2", "3", "4", 
"5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", 
"16", "17", "18", "19", "20", "21", "22", "23", "24")), names = ""), class = "table"))

expect_equal(spe$pg_clusters[1000:1200],
             structure(c(`1000` = 1L, `1001` = 6L, `1002` = 2L, `1003` = 1L, 
`1004` = 1L, `1005` = 1L, `1006` = 1L, `1007` = 2L, `1008` = 1L, 
`1009` = 2L, `1010` = 2L, `1011` = 1L, `1012` = 2L, `1013` = 2L, 
`1014` = 1L, `1015` = 2L, `1016` = 2L, `1017` = 1L, `1018` = 2L, 
`1019` = 1L, `1020` = 1L, `1021` = 2L, `1022` = 2L, `1023` = 1L, 
`1024` = 1L, `1025` = 1L, `1026` = 2L, `1027` = 2L, `1028` = 2L, 
`1029` = 1L, `1030` = 2L, `1031` = 5L, `1032` = 6L, `1033` = 1L, 
`1034` = 1L, `1035` = 5L, `1036` = 1L, `1037` = 1L, `1038` = 1L, 
`1039` = 1L, `1040` = 2L, `1041` = 1L, `1042` = 1L, `1043` = 1L, 
`1044` = 1L, `1045` = 2L, `1046` = 2L, `1047` = 1L, `1048` = 1L, 
`1049` = 1L, `1050` = 6L, `1051` = 1L, `1052` = 1L, `1053` = 1L, 
`1054` = 1L, `1055` = 1L, `1056` = 7L, `1057` = 1L, `1058` = 2L, 
`1059` = 2L, `1060` = 1L, `1061` = 2L, `1062` = 2L, `1063` = 1L, 
`1064` = 1L, `1065` = 7L, `1066` = 1L, `1067` = 1L, `1068` = 1L, 
`1069` = 1L, `1070` = 1L, `1071` = 1L, `1072` = 1L, `1073` = 1L, 
`1074` = 1L, `1075` = 1L, `1076` = 1L, `1077` = 7L, `1078` = 1L, 
`1079` = 1L, `1080` = 13L, `1081` = 1L, `1082` = 1L, `1083` = 1L, 
`1084` = 1L, `1085` = 1L, `1086` = 2L, `1087` = 14L, `1088` = 1L, 
`1089` = 1L, `1090` = 1L, `1091` = 1L, `1092` = 1L, `1093` = 1L, 
`1094` = 5L, `1095` = 1L, `1096` = 7L, `1097` = 1L, `1098` = 1L, 
`1099` = 1L, `1100` = 6L, `1101` = 1L, `1102` = 2L, `1103` = 2L, 
`1104` = 1L, `1105` = 1L, `1106` = 1L, `1107` = 1L, `1108` = 1L, 
`1109` = 1L, `1110` = 5L, `1111` = 13L, `1112` = 2L, `1113` = 2L, 
`1114` = 1L, `1115` = 1L, `1116` = 1L, `1117` = 1L, `1118` = 1L, 
`1119` = 1L, `1120` = 1L, `1121` = 13L, `1122` = 1L, `1123` = 1L, 
`1124` = 1L, `1125` = 1L, `1126` = 2L, `1127` = 1L, `1128` = 7L, 
`1129` = 2L, `1130` = 5L, `1131` = 1L, `1132` = 1L, `1133` = 1L, 
`1134` = 1L, `1135` = 1L, `1136` = 3L, `1137` = 6L, `1138` = 1L, 
`1139` = 5L, `1140` = 2L, `1141` = 2L, `1142` = 1L, `1143` = 1L, 
`1144` = 1L, `1145` = 1L, `1146` = 1L, `1147` = 14L, `1148` = 14L, 
`1149` = 7L, `1150` = 1L, `1151` = 3L, `1152` = 1L, `1153` = 2L, 
`1154` = 13L, `1155` = 1L, `1156` = 2L, `1157` = 2L, `1158` = 1L, 
`1159` = 1L, `1160` = 1L, `1161` = 2L, `1162` = 7L, `1163` = 1L, 
`1164` = 1L, `1165` = 2L, `1166` = 1L, `1167` = 2L, `1168` = 7L, 
`1169` = 1L, `1170` = 5L, `1171` = 1L, `1172` = 2L, `1173` = 1L, 
`1174` = 2L, `1175` = 1L, `1176` = 1L, `1177` = 2L, `1178` = 2L, 
`1179` = 2L, `1180` = 1L, `1181` = 1L, `1182` = 3L, `1183` = 2L, 
`1184` = 1L, `1185` = 1L, `1186` = 1L, `1187` = 1L, `1188` = 1L, 
`1189` = 1L, `1190` = 14L, `1191` = 2L, `1192` = 1L, `1193` = 2L, 
`1194` = 2L, `1195` = 1L, `1196` = 1L, `1197` = 6L, `1198` = 9L, 
`1199` = 14L, `1200` = 6L), levels = c("1", "2", "3", "4", "5", 
"6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", 
"17", "18", "19", "20", "21", "22", "23", "24"), class = "factor"))

expect_equal(table(spe$pg_clusters_corrected), 
             structure(c(`1` = 4103L, `2` = 1545L, `3` = 2684L, `4` = 2372L, 
`5` = 1963L, `6` = 925L, `7` = 3867L, `8` = 684L, `9` = 4598L, 
`10` = 3665L, `11` = 2003L, `12` = 2570L, `13` = 1500L, `14` = 1780L, 
`15` = 1395L, `16` = 600L, `17` = 3025L, `18` = 1214L, `19` = 3426L, 
`20` = 858L, `21` = 871L, `22` = 1057L, `23` = 761L, `24` = 328L
), dim = 24L, dimnames = structure(list(c("1", "2", "3", "4", 
"5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", 
"16", "17", "18", "19", "20", "21", "22", "23", "24")), names = ""), class = "table"))

expect_equal(spe$pg_clusters_corrected[1000:1200],
             structure(c(`1000` = 1L, `1001` = 10L, `1002` = 1L, `1003` = 1L, 
`1004` = 1L, `1005` = 1L, `1006` = 1L, `1007` = 1L, `1008` = 1L, 
`1009` = 5L, `1010` = 12L, `1011` = 1L, `1012` = 1L, `1013` = 1L, 
`1014` = 1L, `1015` = 12L, `1016` = 1L, `1017` = 1L, `1018` = 1L, 
`1019` = 1L, `1020` = 2L, `1021` = 12L, `1022` = 12L, `1023` = 3L, 
`1024` = 1L, `1025` = 7L, `1026` = 1L, `1027` = 1L, `1028` = 1L, 
`1029` = 2L, `1030` = 1L, `1031` = 13L, `1032` = 10L, `1033` = 1L, 
`1034` = 11L, `1035` = 13L, `1036` = 1L, `1037` = 1L, `1038` = 1L, 
`1039` = 1L, `1040` = 1L, `1041` = 1L, `1042` = 1L, `1043` = 2L, 
`1044` = 1L, `1045` = 1L, `1046` = 1L, `1047` = 2L, `1048` = 1L, 
`1049` = 1L, `1050` = 10L, `1051` = 2L, `1052` = 1L, `1053` = 1L, 
`1054` = 1L, `1055` = 1L, `1056` = 13L, `1057` = 1L, `1058` = 12L, 
`1059` = 5L, `1060` = 1L, `1061` = 1L, `1062` = 12L, `1063` = 16L, 
`1064` = 1L, `1065` = 9L, `1066` = 1L, `1067` = 1L, `1068` = 1L, 
`1069` = 1L, `1070` = 1L, `1071` = 1L, `1072` = 11L, `1073` = 3L, 
`1074` = 1L, `1075` = 2L, `1076` = 1L, `1077` = 13L, `1078` = 1L, 
`1079` = 1L, `1080` = 7L, `1081` = 1L, `1082` = 1L, `1083` = 1L, 
`1084` = 1L, `1085` = 1L, `1086` = 1L, `1087` = 9L, `1088` = 1L, 
`1089` = 1L, `1090` = 1L, `1091` = 11L, `1092` = 6L, `1093` = 1L, 
`1094` = 9L, `1095` = 11L, `1096` = 6L, `1097` = 1L, `1098` = 1L, 
`1099` = 11L, `1100` = 10L, `1101` = 1L, `1102` = 1L, `1103` = 12L, 
`1104` = 6L, `1105` = 2L, `1106` = 11L, `1107` = 1L, `1108` = 1L, 
`1109` = 2L, `1110` = 13L, `1111` = 15L, `1112` = 3L, `1113` = 1L, 
`1114` = 1L, `1115` = 1L, `1116` = 1L, `1117` = 1L, `1118` = 1L, 
`1119` = 11L, `1120` = 2L, `1121` = 7L, `1122` = 11L, `1123` = 1L, 
`1124` = 1L, `1125` = 1L, `1126` = 1L, `1127` = 1L, `1128` = 13L, 
`1129` = 1L, `1130` = 9L, `1131` = 1L, `1132` = 1L, `1133` = 1L, 
`1134` = 6L, `1135` = 1L, `1136` = 7L, `1137` = 10L, `1138` = 11L, 
`1139` = 10L, `1140` = 1L, `1141` = 1L, `1142` = 6L, `1143` = 1L, 
`1144` = 1L, `1145` = 1L, `1146` = 1L, `1147` = 11L, `1148` = 11L, 
`1149` = 11L, `1150` = 6L, `1151` = 4L, `1152` = 1L, `1153` = 1L, 
`1154` = 7L, `1155` = 1L, `1156` = 12L, `1157` = 18L, `1158` = 6L, 
`1159` = 1L, `1160` = 1L, `1161` = 1L, `1162` = 11L, `1163` = 1L, 
`1164` = 1L, `1165` = 3L, `1166` = 1L, `1167` = 1L, `1168` = 9L, 
`1169` = 1L, `1170` = 17L, `1171` = 1L, `1172` = 2L, `1173` = 1L, 
`1174` = 1L, `1175` = 2L, `1176` = 6L, `1177` = 1L, `1178` = 1L, 
`1179` = 11L, `1180` = 1L, `1181` = 1L, `1182` = 4L, `1183` = 1L, 
`1184` = 2L, `1185` = 1L, `1186` = 1L, `1187` = 6L, `1188` = 6L, 
`1189` = 1L, `1190` = 9L, `1191` = 1L, `1192` = 2L, `1193` = 5L, 
`1194` = 5L, `1195` = 1L, `1196` = 1L, `1197` = 10L, `1198` = 1L, 
`1199` = 11L, `1200` = 10L), levels = c("1", "2", "3", "4", "5", 
"6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", 
"17", "18", "19", "20", "21", "22", "23", "24"), class = "factor"))

expect_equal(table(spe$nn_clusters), 
             structure(c(`1` = 5444L, `2` = 3603L, `3` = 2930L, `4` = 2135L, 
`5` = 2589L, `6` = 4140L, `7` = 1985L, `8` = 3494L, `9` = 2924L, 
`10` = 2970L, `11` = 3553L, `12` = 1092L, `13` = 4125L, `14` = 2068L, 
`15` = 3181L, `16` = 770L, `17` = 791L), dim = 17L, dimnames = structure(list(
    c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", 
    "12", "13", "14", "15", "16", "17")), names = ""), class = "table"))

expect_equal(spe$nn_clusters[1000:1200],
             structure(c(1L, 5L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 6L, 4L, 1L, 1L, 6L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 1L, 1L, 1L, 5L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 1L, 1L, 6L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 1L, 4L, 1L, 1L, 1L, 1L, 1L, 1L, 11L, 1L, 1L, 1L, 1L, 
1L, 1L, 6L, 1L, 2L, 1L, 1L, 1L, 5L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 6L, 9L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
1L, 1L, 1L, 1L, 1L, 1L, 6L, 1L, 1L, 1L, 1L, 1L, 2L, 4L, 1L, 5L, 
1L, 1L, 1L, 1L, 1L, 1L, 1L, 11L, 11L, 11L, 1L, 2L, 1L, 1L, 4L, 
1L, 1L, 1L, 1L, 1L, 1L, 1L, 12L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
6L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 1L, 1L, 1L, 
1L, 1L, 1L, 1L, 11L, 1L, 1L, 1L, 1L, 1L, 1L, 5L, 6L, 11L, 5L), levels = c("1", 
"2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", 
"14", "15", "16", "17"), class = "factor"))

expect_equal(table(spe$nn_clusters_corrected), 
             structure(c(`1` = 4300L, `2` = 3039L, `3` = 5416L, `4` = 4799L, 
`5` = 4448L, `6` = 2254L, `7` = 4263L, `8` = 5250L, `9` = 3645L, 
`10` = 5016L, `11` = 3574L, `12` = 1598L, `13` = 192L), dim = 13L, dimnames = structure(list(
    c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", 
    "12", "13")), names = ""), class = "table"))

expect_equal(spe$nn_clusters_corrected[1000:1200],
             structure(c(1L, 9L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 3L, 3L, 1L, 1L, 
1L, 1L, 3L, 1L, 1L, 3L, 1L, 2L, 3L, 3L, 6L, 1L, 2L, 1L, 1L, 1L, 
2L, 1L, 10L, 9L, 1L, 1L, 10L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 
1L, 1L, 1L, 1L, 1L, 1L, 9L, 2L, 1L, 1L, 1L, 1L, 10L, 1L, 3L, 
3L, 1L, 3L, 3L, 8L, 1L, 8L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 4L, 1L, 
2L, 1L, 10L, 1L, 1L, 5L, 1L, 1L, 1L, 1L, 1L, 1L, 8L, 1L, 1L, 
1L, 1L, 6L, 1L, 7L, 1L, 5L, 1L, 1L, 1L, 9L, 1L, 1L, 3L, 1L, 2L, 
1L, 1L, 6L, 2L, 10L, 7L, 4L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 
7L, 6L, 1L, 1L, 1L, 4L, 1L, 10L, 1L, 8L, 1L, 1L, 1L, 6L, 1L, 
5L, 9L, 6L, 9L, 1L, 1L, 6L, 1L, 1L, 1L, 1L, 8L, 8L, 8L, 6L, 5L, 
1L, 1L, 7L, 1L, 3L, 4L, 6L, 1L, 1L, 1L, 6L, 1L, 1L, 4L, 1L, 1L, 
8L, 1L, 7L, 1L, 2L, 1L, 1L, 2L, 6L, 1L, 1L, 1L, 1L, 1L, 5L, 1L, 
2L, 1L, 1L, 6L, 1L, 1L, 8L, 1L, 2L, 3L, 3L, 1L, 1L, 9L, 3L, 8L, 
9L), levels = c("1", "2", "3", "4", "5", "6", "7", "8", "9", 
"10", "11", "12", "13"), class = "factor"))

expect_equal(table(spe$som_clusters), 
             structure(c(`1` = 13927L, `2` = 6206L, `3` = 1957L, `4` = 984L, 
`5` = 532L, `6` = 821L, `7` = 322L, `8` = 4285L, `9` = 981L, 
`10` = 9370L, `11` = 973L, `12` = 5862L, `13` = 1574L), dim = 13L, dimnames = structure(list(
    c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", 
    "12", "13")), names = ""), class = "table"))

expect_equal(spe$som_clusters[1000:1200],
             structure(c(13L, 1L, 10L, 10L, 10L, 13L, 10L, 10L, 10L, 10L, 
10L, 13L, 10L, 10L, 13L, 10L, 10L, 10L, 10L, 10L, 13L, 10L, 10L, 
10L, 10L, 10L, 10L, 10L, 10L, 13L, 10L, 8L, 8L, 10L, 10L, 8L, 
13L, 10L, 10L, 10L, 10L, 10L, 10L, 13L, 13L, 10L, 10L, 13L, 13L, 
10L, 1L, 13L, 10L, 10L, 13L, 10L, 13L, 10L, 10L, 10L, 10L, 10L, 
10L, 13L, 10L, 8L, 10L, 10L, 10L, 10L, 10L, 6L, 10L, 10L, 10L, 
13L, 10L, 8L, 10L, 10L, 1L, 13L, 13L, 10L, 10L, 10L, 10L, 2L, 
10L, 10L, 10L, 6L, 13L, 13L, 8L, 10L, 6L, 10L, 10L, 10L, 1L, 
10L, 10L, 10L, 13L, 13L, 6L, 13L, 13L, 10L, 8L, 1L, 10L, 10L, 
10L, 13L, 10L, 10L, 10L, 6L, 13L, 13L, 10L, 10L, 10L, 13L, 10L, 
13L, 8L, 10L, 8L, 13L, 13L, 10L, 13L, 13L, 1L, 8L, 13L, 1L, 10L, 
10L, 13L, 10L, 10L, 13L, 13L, 2L, 10L, 6L, 13L, 1L, 13L, 10L, 
8L, 10L, 10L, 10L, 13L, 10L, 10L, 10L, 6L, 10L, 10L, 10L, 10L, 
10L, 13L, 10L, 8L, 13L, 10L, 10L, 10L, 13L, 13L, 10L, 10L, 10L, 
10L, 10L, 1L, 10L, 13L, 13L, 10L, 13L, 13L, 10L, 8L, 10L, 13L, 
10L, 10L, 10L, 10L, 1L, 10L, 2L, 1L), levels = c("1", "2", "3", 
"4", "5", "6", "7", "8", "9", "10", "11", "12", "13"), class = "factor"))

expect_equal(table(spe$som_clusters_corrected), 
             structure(c(`1` = 5610L, `2` = 3633L, `3` = 1146L, `4` = 109L, 
`5` = 880L, `6` = 950L, `7` = 286L, `8` = 1815L, `9` = 7438L, 
`10` = 8866L, `11` = 1810L, `12` = 2925L, `13` = 9494L, `14` = 1881L, 
`15` = 398L, `16` = 553L), dim = 16L, dimnames = structure(list(
    c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", 
    "12", "13", "14", "15", "16")), names = ""), class = "table"))

expect_equal(spe$som_clusters_corrected[1000:1200],
             structure(c(V33 = 1L, V50 = 13L, V11 = 1L, V32 = 1L, V33 = 1L, 
V32 = 1L, V11 = 1L, V33 = 1L, V33 = 1L, V4 = 2L, V1 = 1L, V32 = 1L, 
V34 = 9L, V11 = 1L, V32 = 1L, V1 = 1L, V33 = 1L, V33 = 1L, V34 = 9L, 
V34 = 9L, V43 = 13L, V1 = 1L, V1 = 1L, V44 = 14L, V11 = 1L, V36 = 9L, 
V14 = 2L, V33 = 1L, V11 = 1L, V43 = 13L, V12 = 1L, V73 = 10L, 
V60 = 13L, V33 = 1L, V24 = 1L, V62 = 10L, V32 = 1L, V33 = 1L, 
V11 = 1L, V13 = 1L, V14 = 2L, V21 = 1L, V32 = 1L, V43 = 13L, 
V21 = 1L, V34 = 9L, V14 = 2L, V32 = 1L, V21 = 1L, V14 = 2L, V30 = 11L, 
V55 = 13L, V21 = 1L, V11 = 1L, V32 = 1L, V32 = 1L, V26 = 10L, 
V11 = 1L, V1 = 1L, V14 = 2L, V14 = 2L, V34 = 9L, V1 = 1L, V53 = 10L, 
V14 = 2L, V83 = 10L, V33 = 1L, V11 = 1L, V33 = 1L, V32 = 1L, 
V33 = 1L, V24 = 1L, V24 = 1L, V45 = 9L, V33 = 1L, V55 = 13L, 
V21 = 1L, V73 = 10L, V34 = 9L, V21 = 1L, V59 = 13L, V21 = 1L, 
V32 = 1L, V14 = 2L, V24 = 1L, V33 = 1L, V33 = 1L, V93 = 10L, 
V33 = 1L, V33 = 1L, V14 = 2L, V24 = 1L, V42 = 8L, V32 = 1L, V78 = 13L, 
V11 = 1L, V31 = 8L, V13 = 1L, V33 = 1L, V24 = 1L, V30 = 11L, 
V33 = 1L, V12 = 1L, V12 = 1L, V21 = 1L, V43 = 13L, V24 = 1L, 
V13 = 1L, V32 = 1L, V34 = 9L, V73 = 10L, V90 = 13L, V34 = 9L, 
V11 = 1L, V33 = 1L, V32 = 1L, V21 = 1L, V24 = 1L, V21 = 1L, V24 = 1L, 
V55 = 13L, V55 = 13L, V32 = 1L, V21 = 1L, V21 = 1L, V32 = 1L, 
V34 = 9L, V21 = 1L, V73 = 10L, V14 = 2L, V99 = 10L, V32 = 1L, 
V32 = 1L, V13 = 1L, V42 = 8L, V32 = 1L, V36 = 9L, V60 = 13L, 
V32 = 1L, V65 = 11L, V14 = 2L, V11 = 1L, V69 = 13L, V34 = 9L, 
V32 = 1L, V32 = 1L, V32 = 1L, V44 = 14L, V44 = 14L, V44 = 14L, 
V42 = 8L, V29 = 9L, V32 = 1L, V14 = 2L, V100 = 13L, V13 = 1L, 
V1 = 1L, V34 = 9L, V42 = 8L, V33 = 1L, V21 = 1L, V11 = 1L, V31 = 8L, 
V32 = 1L, V11 = 1L, V25 = 9L, V32 = 1L, V33 = 1L, V99 = 10L, 
V13 = 1L, V78 = 13L, V32 = 1L, V34 = 9L, V13 = 1L, V33 = 1L, 
V55 = 13L, V42 = 8L, V34 = 9L, V13 = 1L, V24 = 1L, V11 = 1L, 
V33 = 1L, V29 = 9L, V33 = 1L, V55 = 13L, V21 = 1L, V11 = 1L, 
V42 = 8L, V21 = 1L, V13 = 1L, V99 = 10L, V33 = 1L, V32 = 1L, 
V4 = 2L, V5 = 2L, V33 = 1L, V33 = 1L, V39 = 11L, V5 = 2L, V44 = 14L, 
V39 = 11L), levels = c("1", "2", "3", "4", "5", "6", "7", "8", 
"9", "10", "11", "12", "13", "14", "15", "16"), class = "factor"))

expect_equal(table(spe$cell_labels), 
             structure(c(Bcell = 780L, BnTcell = 1702L, CD4 = 688L, CD8 = 822L, 
Myeloid = 1750L, Neutrophil = 134L, Plasma_cell = 716L, Stroma = 442L, 
Treg = 361L, Tumor = 5999L, unlabeled = 34400L), dim = 11L, dimnames = structure(list(
    c("Bcell", "BnTcell", "CD4", "CD8", "Myeloid", "Neutrophil", 
    "Plasma_cell", "Stroma", "Treg", "Tumor", "unlabeled")), names = ""), class = "table"))

expect_equal(spe$cell_labels[1000:1200],
             c(Patient1_001_1001 = "unlabeled", Patient1_001_1002 = "unlabeled", 
Patient1_001_1003 = "unlabeled", Patient1_001_1004 = "unlabeled", 
Patient1_001_1005 = "unlabeled", Patient1_001_1006 = "unlabeled", 
Patient1_001_1007 = "unlabeled", Patient1_001_1008 = "unlabeled", 
Patient1_001_1009 = "unlabeled", Patient1_001_1010 = "unlabeled", 
Patient1_001_1011 = "unlabeled", Patient1_001_1012 = "unlabeled", 
Patient1_001_1013 = "unlabeled", Patient1_001_1014 = "unlabeled", 
Patient1_001_1015 = "unlabeled", Patient1_001_1016 = "unlabeled", 
Patient1_001_1017 = "unlabeled", Patient1_001_1018 = "unlabeled", 
Patient1_001_1019 = "unlabeled", Patient1_001_1020 = "unlabeled", 
Patient1_001_1021 = "unlabeled", Patient1_001_1022 = "unlabeled", 
Patient1_001_1023 = "unlabeled", Patient1_001_1024 = "unlabeled", 
Patient1_001_1025 = "unlabeled", Patient1_001_1026 = "unlabeled", 
Patient1_001_1027 = "unlabeled", Patient1_001_1028 = "unlabeled", 
Patient1_001_1029 = "unlabeled", Patient1_001_1030 = "unlabeled", 
Patient1_001_1031 = "unlabeled", Patient1_001_1032 = "unlabeled", 
Patient1_001_1033 = "unlabeled", Patient1_001_1034 = "unlabeled", 
Patient1_001_1035 = "unlabeled", Patient1_001_1036 = "unlabeled", 
Patient1_001_1037 = "unlabeled", Patient1_001_1038 = "unlabeled", 
Patient1_001_1039 = "unlabeled", Patient1_001_1040 = "unlabeled", 
Patient1_001_1041 = "unlabeled", Patient1_001_1042 = "unlabeled", 
Patient1_001_1043 = "unlabeled", Patient1_001_1044 = "unlabeled", 
Patient1_001_1045 = "unlabeled", Patient1_001_1046 = "unlabeled", 
Patient1_001_1047 = "unlabeled", Patient1_001_1048 = "unlabeled", 
Patient1_001_1049 = "unlabeled", Patient1_001_1050 = "unlabeled", 
Patient1_001_1051 = "unlabeled", Patient1_001_1052 = "unlabeled", 
Patient1_001_1053 = "unlabeled", Patient1_001_1054 = "unlabeled", 
Patient1_001_1055 = "unlabeled", Patient1_001_1056 = "unlabeled", 
Patient1_001_1057 = "unlabeled", Patient1_001_1058 = "unlabeled", 
Patient1_001_1059 = "unlabeled", Patient1_001_1060 = "unlabeled", 
Patient1_001_1061 = "unlabeled", Patient1_001_1062 = "unlabeled", 
Patient1_001_1063 = "unlabeled", Patient1_001_1064 = "unlabeled", 
Patient1_001_1065 = "unlabeled", Patient1_001_1066 = "unlabeled", 
Patient1_001_1067 = "unlabeled", Patient1_001_1068 = "unlabeled", 
Patient1_001_1069 = "unlabeled", Patient1_001_1070 = "unlabeled", 
Patient1_001_1071 = "unlabeled", Patient1_001_1072 = "unlabeled", 
Patient1_001_1073 = "unlabeled", Patient1_001_1074 = "unlabeled", 
Patient1_001_1075 = "unlabeled", Patient1_001_1076 = "unlabeled", 
Patient1_001_1077 = "unlabeled", Patient1_001_1078 = "unlabeled", 
Patient1_001_1079 = "unlabeled", Patient1_001_1080 = "unlabeled", 
Patient1_001_1081 = "unlabeled", Patient1_001_1082 = "unlabeled", 
Patient1_001_1083 = "unlabeled", Patient1_001_1084 = "unlabeled", 
Patient1_001_1085 = "unlabeled", Patient1_001_1086 = "unlabeled", 
Patient1_001_1087 = "unlabeled", Patient1_001_1088 = "unlabeled", 
Patient1_001_1089 = "unlabeled", Patient1_001_1090 = "unlabeled", 
Patient1_001_1091 = "unlabeled", Patient1_001_1092 = "unlabeled", 
Patient1_001_1093 = "unlabeled", Patient1_001_1094 = "unlabeled", 
Patient1_001_1095 = "unlabeled", Patient1_001_1096 = "unlabeled", 
Patient1_001_1097 = "unlabeled", Patient1_001_1098 = "unlabeled", 
Patient1_001_1099 = "unlabeled", Patient1_001_1100 = "unlabeled", 
Patient1_001_1101 = "unlabeled", Patient1_001_1102 = "unlabeled", 
Patient1_001_1103 = "unlabeled", Patient1_001_1104 = "unlabeled", 
Patient1_001_1105 = "unlabeled", Patient1_001_1106 = "unlabeled", 
Patient1_001_1107 = "unlabeled", Patient1_001_1108 = "unlabeled", 
Patient1_001_1109 = "unlabeled", Patient1_001_1110 = "unlabeled", 
Patient1_001_1111 = "unlabeled", Patient1_001_1112 = "unlabeled", 
Patient1_001_1113 = "unlabeled", Patient1_001_1114 = "unlabeled", 
Patient1_001_1115 = "unlabeled", Patient1_001_1116 = "unlabeled", 
Patient1_001_1117 = "unlabeled", Patient1_001_1118 = "unlabeled", 
Patient1_001_1119 = "unlabeled", Patient1_001_1120 = "unlabeled", 
Patient1_001_1121 = "unlabeled", Patient1_001_1122 = "unlabeled", 
Patient1_001_1123 = "unlabeled", Patient1_001_1124 = "unlabeled", 
Patient1_001_1125 = "unlabeled", Patient1_001_1126 = "unlabeled", 
Patient1_001_1127 = "unlabeled", Patient1_001_1128 = "unlabeled", 
Patient1_001_1129 = "unlabeled", Patient1_001_1130 = "unlabeled", 
Patient1_001_1131 = "unlabeled", Patient1_001_1132 = "unlabeled", 
Patient1_001_1133 = "unlabeled", Patient1_001_1134 = "unlabeled", 
Patient1_001_1135 = "unlabeled", Patient1_001_1136 = "unlabeled", 
Patient1_001_1137 = "unlabeled", Patient1_001_1138 = "unlabeled", 
Patient1_001_1139 = "unlabeled", Patient1_001_1140 = "unlabeled", 
Patient1_001_1141 = "unlabeled", Patient1_001_1142 = "unlabeled", 
Patient1_001_1143 = "unlabeled", Patient1_001_1144 = "unlabeled", 
Patient1_001_1145 = "unlabeled", Patient1_001_1146 = "unlabeled", 
Patient1_001_1147 = "unlabeled", Patient1_001_1148 = "unlabeled", 
Patient1_001_1149 = "unlabeled", Patient1_001_1150 = "unlabeled", 
Patient1_001_1151 = "unlabeled", Patient1_001_1152 = "Stroma", 
Patient1_001_1153 = "unlabeled", Patient1_001_1154 = "unlabeled", 
Patient1_001_1155 = "unlabeled", Patient1_001_1156 = "unlabeled", 
Patient1_001_1157 = "unlabeled", Patient1_001_1158 = "unlabeled", 
Patient1_001_1159 = "unlabeled", Patient1_001_1160 = "unlabeled", 
Patient1_001_1161 = "unlabeled", Patient1_001_1162 = "unlabeled", 
Patient1_001_1163 = "unlabeled", Patient1_001_1164 = "unlabeled", 
Patient1_001_1165 = "unlabeled", Patient1_001_1166 = "unlabeled", 
Patient1_001_1167 = "unlabeled", Patient1_001_1168 = "unlabeled", 
Patient1_001_1169 = "unlabeled", Patient1_001_1170 = "unlabeled", 
Patient1_001_1171 = "unlabeled", Patient1_001_1172 = "unlabeled", 
Patient1_001_1173 = "unlabeled", Patient1_001_1174 = "unlabeled", 
Patient1_001_1175 = "unlabeled", Patient1_001_1176 = "unlabeled", 
Patient1_001_1177 = "unlabeled", Patient1_001_1178 = "unlabeled", 
Patient1_001_1179 = "unlabeled", Patient1_001_1180 = "unlabeled", 
Patient1_001_1181 = "unlabeled", Patient1_001_1182 = "unlabeled", 
Patient1_001_1183 = "Stroma", Patient1_001_1184 = "unlabeled", 
Patient1_001_1185 = "unlabeled", Patient1_001_1186 = "unlabeled", 
Patient1_001_1187 = "unlabeled", Patient1_001_1188 = "unlabeled", 
Patient1_001_1189 = "unlabeled", Patient1_001_1190 = "unlabeled", 
Patient1_001_1191 = "unlabeled", Patient1_001_1192 = "unlabeled", 
Patient1_001_1193 = "unlabeled", Patient1_001_1194 = "unlabeled", 
Patient1_001_1195 = "unlabeled", Patient1_001_1196 = "unlabeled", 
Patient1_001_1197 = "unlabeled", Patient1_001_1198 = "unlabeled", 
Patient1_001_1199 = "unlabeled", Patient1_001_1200 = "unlabeled", 
Patient1_001_1201 = "unlabeled"))

expect_equal(table(spe$celltype), 
             structure(c(Bcell = 1595L, BnTcell = 2674L, CD4 = 3845L, CD8 = 3504L, 
Myeloid = 8020L, Neutrophil = 681L, Plasma_cell = 3397L, Stroma = 5306L, 
Treg = 1515L, Tumor = 16625L, undefined = 632L), dim = 11L, dimnames = structure(list(
    c("Bcell", "BnTcell", "CD4", "CD8", "Myeloid", "Neutrophil", 
    "Plasma_cell", "Stroma", "Treg", "Tumor", "undefined")), names = ""), class = "table"))

expect_equal(spe$celltype[1000:1200],
             c(Patient1_001_1001 = "Tumor", Patient1_001_1002 = "Plasma_cell", 
Patient1_001_1003 = "Tumor", Patient1_001_1004 = "Tumor", Patient1_001_1005 = "Tumor", 
Patient1_001_1006 = "Tumor", Patient1_001_1007 = "Tumor", Patient1_001_1008 = "Tumor", 
Patient1_001_1009 = "Tumor", Patient1_001_1010 = "Tumor", Patient1_001_1011 = "Tumor", 
Patient1_001_1012 = "Tumor", Patient1_001_1013 = "Tumor", Patient1_001_1014 = "Tumor", 
Patient1_001_1015 = "Tumor", Patient1_001_1016 = "Tumor", Patient1_001_1017 = "Tumor", 
Patient1_001_1018 = "Tumor", Patient1_001_1019 = "Tumor", Patient1_001_1020 = "Tumor", 
Patient1_001_1021 = "Tumor", Patient1_001_1022 = "Tumor", Patient1_001_1023 = "Tumor", 
Patient1_001_1024 = "Tumor", Patient1_001_1025 = "Tumor", Patient1_001_1026 = "Tumor", 
Patient1_001_1027 = "Tumor", Patient1_001_1028 = "Tumor", Patient1_001_1029 = "Tumor", 
Patient1_001_1030 = "Tumor", Patient1_001_1031 = "Tumor", Patient1_001_1032 = "Treg", 
Patient1_001_1033 = "Plasma_cell", Patient1_001_1034 = "Tumor", 
Patient1_001_1035 = "Tumor", Patient1_001_1036 = "Treg", Patient1_001_1037 = "Tumor", 
Patient1_001_1038 = "Tumor", Patient1_001_1039 = "Tumor", Patient1_001_1040 = "Tumor", 
Patient1_001_1041 = "Tumor", Patient1_001_1042 = "Tumor", Patient1_001_1043 = "Tumor", 
Patient1_001_1044 = "CD4", Patient1_001_1045 = "Tumor", Patient1_001_1046 = "Tumor", 
Patient1_001_1047 = "Tumor", Patient1_001_1048 = "Tumor", Patient1_001_1049 = "Tumor", 
Patient1_001_1050 = "Tumor", Patient1_001_1051 = "Plasma_cell", 
Patient1_001_1052 = "Tumor", Patient1_001_1053 = "Tumor", Patient1_001_1054 = "Tumor", 
Patient1_001_1055 = "Tumor", Patient1_001_1056 = "Tumor", Patient1_001_1057 = "CD4", 
Patient1_001_1058 = "Tumor", Patient1_001_1059 = "Tumor", Patient1_001_1060 = "Tumor", 
Patient1_001_1061 = "Tumor", Patient1_001_1062 = "Tumor", Patient1_001_1063 = "Tumor", 
Patient1_001_1064 = "Tumor", Patient1_001_1065 = "Tumor", Patient1_001_1066 = "CD8", 
Patient1_001_1067 = "Tumor", Patient1_001_1068 = "Tumor", Patient1_001_1069 = "Tumor", 
Patient1_001_1070 = "Tumor", Patient1_001_1071 = "Tumor", Patient1_001_1072 = "Treg", 
Patient1_001_1073 = "Tumor", Patient1_001_1074 = "Tumor", Patient1_001_1075 = "Tumor", 
Patient1_001_1076 = "Tumor", Patient1_001_1077 = "Tumor", Patient1_001_1078 = "Treg", 
Patient1_001_1079 = "Tumor", Patient1_001_1080 = "Tumor", Patient1_001_1081 = "CD4", 
Patient1_001_1082 = "Tumor", Patient1_001_1083 = "Tumor", Patient1_001_1084 = "Tumor", 
Patient1_001_1085 = "Tumor", Patient1_001_1086 = "Tumor", Patient1_001_1087 = "Tumor", 
Patient1_001_1088 = "CD8", Patient1_001_1089 = "Tumor", Patient1_001_1090 = "Tumor", 
Patient1_001_1091 = "Tumor", Patient1_001_1092 = "CD8", Patient1_001_1093 = "Tumor", 
Patient1_001_1094 = "Tumor", Patient1_001_1095 = "Myeloid", Patient1_001_1096 = "Tumor", 
Patient1_001_1097 = "Stroma", Patient1_001_1098 = "Tumor", Patient1_001_1099 = "Tumor", 
Patient1_001_1100 = "Tumor", Patient1_001_1101 = "Plasma_cell", 
Patient1_001_1102 = "Tumor", Patient1_001_1103 = "Tumor", Patient1_001_1104 = "Tumor", 
Patient1_001_1105 = "Tumor", Patient1_001_1106 = "Tumor", Patient1_001_1107 = "Tumor", 
Patient1_001_1108 = "Tumor", Patient1_001_1109 = "Tumor", Patient1_001_1110 = "Tumor", 
Patient1_001_1111 = "Treg", Patient1_001_1112 = "Myeloid", Patient1_001_1113 = "Tumor", 
Patient1_001_1114 = "Tumor", Patient1_001_1115 = "Tumor", Patient1_001_1116 = "Tumor", 
Patient1_001_1117 = "Tumor", Patient1_001_1118 = "Tumor", Patient1_001_1119 = "Tumor", 
Patient1_001_1120 = "CD8", Patient1_001_1121 = "Tumor", Patient1_001_1122 = "Tumor", 
Patient1_001_1123 = "Tumor", Patient1_001_1124 = "Tumor", Patient1_001_1125 = "Tumor", 
Patient1_001_1126 = "Tumor", Patient1_001_1127 = "Tumor", Patient1_001_1128 = "Tumor", 
Patient1_001_1129 = "Treg", Patient1_001_1130 = "Tumor", Patient1_001_1131 = "CD8", 
Patient1_001_1132 = "Tumor", Patient1_001_1133 = "Tumor", Patient1_001_1134 = "Tumor", 
Patient1_001_1135 = "Tumor", Patient1_001_1136 = "Tumor", Patient1_001_1137 = "Tumor", 
Patient1_001_1138 = "Plasma_cell", Patient1_001_1139 = "Tumor", 
Patient1_001_1140 = "Stroma", Patient1_001_1141 = "Tumor", Patient1_001_1142 = "Tumor", 
Patient1_001_1143 = "Tumor", Patient1_001_1144 = "Tumor", Patient1_001_1145 = "Tumor", 
Patient1_001_1146 = "Tumor", Patient1_001_1147 = "Tumor", Patient1_001_1148 = "CD8", 
Patient1_001_1149 = "Tumor", Patient1_001_1150 = "Treg", Patient1_001_1151 = "Tumor", 
Patient1_001_1152 = "Stroma", Patient1_001_1153 = "Tumor", Patient1_001_1154 = "Tumor", 
Patient1_001_1155 = "Myeloid", Patient1_001_1156 = "Tumor", Patient1_001_1157 = "Tumor", 
Patient1_001_1158 = "Tumor", Patient1_001_1159 = "Tumor", Patient1_001_1160 = "Tumor", 
Patient1_001_1161 = "Tumor", Patient1_001_1162 = "Tumor", Patient1_001_1163 = "CD4", 
Patient1_001_1164 = "Tumor", Patient1_001_1165 = "Tumor", Patient1_001_1166 = "Tumor", 
Patient1_001_1167 = "Tumor", Patient1_001_1168 = "Tumor", Patient1_001_1169 = "CD8", 
Patient1_001_1170 = "Tumor", Patient1_001_1171 = "Myeloid", Patient1_001_1172 = "Tumor", 
Patient1_001_1173 = "Tumor", Patient1_001_1174 = "Tumor", Patient1_001_1175 = "Tumor", 
Patient1_001_1176 = "Tumor", Patient1_001_1177 = "Tumor", Patient1_001_1178 = "Tumor", 
Patient1_001_1179 = "Tumor", Patient1_001_1180 = "Tumor", Patient1_001_1181 = "Tumor", 
Patient1_001_1182 = "Tumor", Patient1_001_1183 = "Stroma", Patient1_001_1184 = "Tumor", 
Patient1_001_1185 = "Tumor", Patient1_001_1186 = "Tumor", Patient1_001_1187 = "Tumor", 
Patient1_001_1188 = "Tumor", Patient1_001_1189 = "Tumor", Patient1_001_1190 = "Tumor", 
Patient1_001_1191 = "Plasma_cell", Patient1_001_1192 = "Tumor", 
Patient1_001_1193 = "Tumor", Patient1_001_1194 = "Tumor", Patient1_001_1195 = "Tumor", 
Patient1_001_1196 = "Tumor", Patient1_001_1197 = "Tumor", Patient1_001_1198 = "Plasma_cell", 
Patient1_001_1199 = "Tumor", Patient1_001_1200 = "CD8", Patient1_001_1201 = "Plasma_cell"
))
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}
sessionInfo()
```
</details>
